refactor: split task_executor.rs into pipeline modules (#772)
Split 2,927-line task_executor.rs (violating U-16 800-line limit) into
focused pipeline modules under task_executor/:

- mod.rs (~675 lines): TaskContext, run_task thin orchestrator,
  run_test_gate, TaskTargetDir, restricted_tools
- agent_review.rs (~455 lines): run_agent_review, jaccard_word_similarity,
  normalize_issues + unit tests
- implement_pipeline.rs (~765 lines): run_implement_phase (prompt
  construction + implement block), ImplementOutcome, compute_backoff_ms,
  prepend_constitution, parse_implementation_outcome + unit tests
- review_loop.rs (~448 lines): run_review_loop (external review bot loop)
- triage_pipeline.rs (~287 lines): run_triage_plan_pipeline,
  run_plan_for_prompt, run_rebase_turn
- turn_lifecycle.rs (~281 lines): run_turn_lifecycle
- helpers.rs: + run_agent_streaming, persist_artifact

Part 2: migrate 4 inline prompt templates to prompts.rs:
- plan_for_prompt_task(), validation_retry_prompt(),
  test_gate_failure_prompt(), agent_review_capability_note()

No behavior changes. All 721+ tests pass.
