--- crates/harness-server/src/task_executor.rs
@@ -1,2927 +1,0 @@
-pub(crate) mod conflict_resolver;
-pub(crate) mod helpers;
-pub(crate) mod pr_detection;
-
-use crate::task_runner::{
-    mutate_and_persist, CreateTaskRequest, RoundResult, TaskId, TaskStatus, TaskStore,
-};
-use anyhow::Context;
-use harness_core::agent::{
-    AgentEvent, AgentRequest, AgentResponse, CodeAgent, StreamItem, TurnRequest,
-};
-use harness_core::config::agents::CapabilityProfile;
-use harness_core::error::HarnessError;
-use harness_core::interceptor::ToolUseEvent;
-use harness_core::tool_isolation::validate_tool_usage;
-use harness_core::types::{
-    ContextItem, Decision, Event, ExecutionPhase, Item, SessionId, ThreadId, TokenUsage, TurnId,
-    TurnStatus,
-};
-use harness_core::{config::project::load_project_config, lang_detect, prompts};
-use harness_protocol::{notifications::Notification, notifications::RpcNotification};
-use std::collections::{HashMap, HashSet};
-
-/// Extract tool list from a capability profile, returning an error if the
-/// profile unexpectedly returns `None` (which means Full/unrestricted).
-/// A misconfigured profile causes a hard failure rather than silent degradation,
-/// per U-23 (no silent capability downgrade).
-fn restricted_tools(profile: CapabilityProfile) -> anyhow::Result<Vec<String>> {
-    profile.tools().ok_or_else(|| {
-        anyhow::anyhow!(
-            "capability profile {:?} returned None from tools() — misconfiguration",
-            profile
-        )
-    })
-}
-use helpers::{
-    collect_context_items, detect_modified_files, emit_runtime_notification,
-    inject_skills_into_prompt, mark_turn_failed, matched_skills_for_prompt, persist_runtime_thread,
-    process_stream_item, run_on_error, run_post_execute, run_post_tool_use, run_pre_execute,
-    truncate_validation_error, update_status,
-};
-#[cfg(test)]
-use pr_detection::{
-    build_fix_ci_prompt, parse_harness_mention_command, HarnessMentionCommand, PromptBuilder,
-};
-use pr_detection::{detect_repo_slug, find_existing_pr_for_issue};
-use std::path::{Path, PathBuf};
-use std::sync::Arc;
-use tokio::sync::{mpsc, RwLock};
-
-/// RAII guard that removes the per-task Cargo target directory on drop.
-/// This ensures cleanup regardless of how `run_task` exits (success, error,
-/// or timeout), preventing disk exhaustion from accumulated build artifacts.
-struct TaskTargetDir(PathBuf);
-
-impl Drop for TaskTargetDir {
-    fn drop(&mut self) {
-        if self.0.exists() {
-            if let Err(e) = std::fs::remove_dir_all(&self.0) {
-                tracing::warn!(
-                    path = %self.0.display(),
-                    "failed to remove per-task cargo target dir: {e}"
-                );
-            }
-        }
-    }
-}
-use tokio::process::Command as TokioCommand;
-use tokio::time::{sleep, timeout, Duration, Instant};
-
-/// Run the project's test commands as a hard gate before accepting LGTM.
-///
-/// When `custom_cmds` is non-empty (from `validation.pre_push` in project config),
-/// those commands are run in order instead of language-detected defaults.
-/// When `custom_cmds` is empty, falls back to language detection.
-///
-/// Returns `Ok(())` when all commands pass or when no test command is detectable
-/// (soft degradation — unknown project type skips rather than hard-fails).
-///
-/// Returns `Err(output)` containing stdout/stderr of the first failing command.
-async fn run_test_gate(
-    project_root: &Path,
-    custom_cmds: &[String],
-    timeout_secs: u64,
-    extra_env: &HashMap<String, String>,
-) -> Result<(), String> {
-    // Prefer explicitly configured pre_push commands; fall back to language detection.
-    let cmds: Vec<String> = if !custom_cmds.is_empty() {
-        // Issue 1 fix: validate every custom command against the safety allowlist
-        // before executing. Malicious repos could supply shell-injection payloads
-        // via `.harness/config.toml` validation.pre_push.
-        for cmd in custom_cmds {
-            if let Err(e) = crate::post_validator::validate_command_safety(cmd) {
-                return Err(format!("test gate: command rejected by safety check: {e}"));
-            }
-        }
-        custom_cmds.to_vec()
-    } else {
-        match lang_detect::primary_test_command(project_root) {
-            Some(cmd) => vec![cmd],
-            None => {
-                tracing::info!(
-                    project = %project_root.display(),
-                    "test gate: no test command detected for project, skipping"
-                );
-                return Ok(());
-            }
-        }
-    };
-
-    for cmd in &cmds {
-        tracing::info!(cmd = %cmd, "test gate: running tests before accepting LGTM");
-
-        let child = match TokioCommand::new("sh")
-            .args(["-c", cmd])
-            .current_dir(project_root)
-            // Issue 3 fix: inherit the per-task CARGO_TARGET_DIR so parallel
-            // Rust tasks do not contend on the same build directory (issue #488).
-            .envs(extra_env)
-            .stdin(std::process::Stdio::null())
-            .stdout(std::process::Stdio::piped())
-            .stderr(std::process::Stdio::piped())
-            .kill_on_drop(true)
-            .spawn()
-        {
-            Ok(c) => c,
-            Err(e) => return Err(format!("test gate: failed to spawn `{cmd}`: {e}")),
-        };
-
-        match timeout(Duration::from_secs(timeout_secs), child.wait_with_output()).await {
-            Ok(Ok(out)) if out.status.success() => {
-                tracing::info!(cmd = %cmd, "test gate: tests passed");
-            }
-            Ok(Ok(out)) => {
-                let stdout = String::from_utf8_lossy(&out.stdout);
-                let stderr = String::from_utf8_lossy(&out.stderr);
-                let code = out.status.code().unwrap_or(-1);
-                return Err(format!(
-                    "Test gate failed (exit {code})\nstdout:\n{stdout}\nstderr:\n{stderr}"
-                ));
-            }
-            Ok(Err(e)) => return Err(format!("test gate: `{cmd}` failed to wait: {e}")),
-            Err(_) => {
-                return Err(format!(
-                    "Test gate timed out after {timeout_secs}s (command: `{cmd}`)"
-                ))
-            }
-        }
-    }
-
-    Ok(())
-}
-
-pub(crate) async fn run_turn_lifecycle(
-    server: Arc<crate::server::HarnessServer>,
-    thread_db: Option<crate::thread_db::ThreadDb>,
-    notify_tx: Option<crate::notify::NotifySender>,
-    notification_tx: tokio::sync::broadcast::Sender<RpcNotification>,
-    thread_id: ThreadId,
-    turn_id: TurnId,
-    prompt: String,
-    agent_name: String,
-) {
-    let Some(project_root) = server
-        .thread_manager
-        .get_thread(&thread_id)
-        .map(|thread| thread.project_root)
-    else {
-        tracing::warn!(
-            "run_turn_lifecycle skipped because thread {} no longer exists",
-            thread_id
-        );
-        return;
-    };
-
-    let Some(agent) = server.agent_registry.get(&agent_name) else {
-        let msg = format!("agent `{agent_name}` not found in registry");
-        if let Err(e) = server.thread_manager.add_item(
-            &thread_id,
-            &turn_id,
-            harness_core::types::Item::Error {
-                code: -1,
-                message: msg.clone(),
-            },
-        ) {
-            tracing::warn!("failed to add agent-not-found error item: {e}");
-        }
-        mark_turn_failed(
-            &server,
-            &thread_db,
-            &notify_tx,
-            &notification_tx,
-            &thread_id,
-            &turn_id,
-            msg,
-        )
-        .await;
-        return;
-    };
-
-    // RAII guard: ensures the adapter is deregistered when the turn scope exits,
-    // even if the task is cancelled before reaching the end of this function.
-    struct AdapterGuard {
-        server: Arc<crate::server::HarnessServer>,
-        turn_id: TurnId,
-    }
-    impl Drop for AdapterGuard {
-        fn drop(&mut self) {
-            self.server
-                .thread_manager
-                .deregister_active_adapter(&self.turn_id);
-        }
-    }
-
-    // Get the adapter for this agent, if any.
-    let adapter_opt = server.agent_registry.get_adapter(&agent_name);
-
-    // Register as live adapter (RAII guard for cleanup on turn exit).
-    // When an adapter is available, execution goes through adapter.start_turn()
-    // (see below), so the adapter's stdin is initialized before any
-    // steer/respond_approval call arrives.
-    let _adapter_guard = adapter_opt.as_ref().map(|adapter_arc| {
-        server
-            .thread_manager
-            .register_active_adapter(&turn_id, adapter_arc.clone());
-        AdapterGuard {
-            server: server.clone(),
-            turn_id: turn_id.clone(),
-        }
-    });
-
-    let stall_timeout = Duration::from_secs(server.config.concurrency.stall_timeout_secs);
-    let (stream_tx, mut stream_rx) = mpsc::channel(128);
-
-    // Only Codex uses the adapter for turn execution (initializes stdin so that
-    // subsequent steer/respond_approval calls can find a live process).
-    // Other adapters (e.g., ClaudeAdapter) exist for interrupt/steer only and must
-    // not override the configured CodeAgent execution path — ClaudeCodeAgent carries
-    // config-level settings (model, sandbox) that ClaudeAdapter does not replicate.
-    let execution_adapter = adapter_opt
-        .as_ref()
-        .filter(|a| a.name() == "codex")
-        .cloned();
-    let mut execution: std::pin::Pin<
-        Box<dyn std::future::Future<Output = harness_core::error::Result<()>> + Send>,
-    > = if let Some(adapter_arc) = execution_adapter {
-        let (event_tx, mut event_rx) = mpsc::channel::<AgentEvent>(128);
-        // Move stream_tx into the bridge task so dropping it closes stream_rx.
-        let bridge_tx = stream_tx;
-        tokio::spawn(async move {
-            let mut output_buf = String::new();
-            while let Some(event) = event_rx.recv().await {
-                let maybe_item: Option<StreamItem> = match event {
-                    AgentEvent::MessageDelta { ref text } => {
-                        output_buf.push_str(text);
-                        Some(StreamItem::MessageDelta { text: text.clone() })
-                    }
-                    AgentEvent::ApprovalRequest { id, command } => {
-                        Some(StreamItem::ApprovalRequest { id, command })
-                    }
-                    AgentEvent::Error { message } => Some(StreamItem::Error { message }),
-                    AgentEvent::TurnCompleted { output } => {
-                        let content = if output.is_empty() {
-                            std::mem::take(&mut output_buf)
-                        } else {
-                            output
-                        };
-                        Some(StreamItem::ItemCompleted {
-                            item: harness_core::types::Item::AgentReasoning { content },
-                        })
-                    }
-                    _ => None,
-                };
-                if let Some(item) = maybe_item {
-                    if bridge_tx.send(item).await.is_err() {
-                        return;
-                    }
-                }
-            }
-            // event_rx closed → adapter done; dropping bridge_tx closes stream_rx.
-        });
-        let turn_req = TurnRequest {
-            prompt,
-            project_root,
-            model: None,
-            allowed_tools: vec![],
-            context: vec![],
-            timeout_secs: None,
-        };
-        Box::pin(async move { adapter_arc.start_turn(turn_req, event_tx).await })
-    } else {
-        let req = AgentRequest {
-            prompt,
-            project_root,
-            ..Default::default()
-        };
-        Box::pin(agent.execute_stream(req, stream_tx))
-    };
-    let mut stream_closed = false;
-    let mut execution_result: Option<harness_core::error::Result<()>> = None;
-    let mut last_activity = Instant::now();
-
-    'outer: while execution_result.is_none() || !stream_closed {
-        tokio::select! {
-            result = &mut execution, if execution_result.is_none() => {
-                execution_result = Some(result);
-            }
-            incoming = stream_rx.recv(), if !stream_closed => {
-                match incoming {
-                    Some(item) => {
-                        last_activity = Instant::now();
-                        process_stream_item(
-                            &server,
-                            &thread_db,
-                            &notify_tx,
-                            &notification_tx,
-                            &thread_id,
-                            &turn_id,
-                            item,
-                        ).await;
-                    }
-                    None => {
-                        stream_closed = true;
-                    }
-                }
-            }
-            _ = tokio::time::sleep_until(last_activity + stall_timeout) => {
-                let elapsed = last_activity.elapsed();
-                tracing::warn!(
-                    thread_id = %thread_id,
-                    turn_id = %turn_id,
-                    elapsed_secs = elapsed.as_secs(),
-                    "agent stream stall detected; no output for {}s",
-                    stall_timeout.as_secs()
-                );
-                // Store the stall reason as the execution result so the Err branch
-                // below appends a stall-specific Item::Error before marking failed.
-                execution_result = Some(Err(HarnessError::AgentExecution(format!(
-                    "Agent stream stalled: no output for {}s",
-                    stall_timeout.as_secs()
-                ))));
-                break 'outer;
-            }
-        }
-    }
-
-    match execution_result.unwrap_or_else(|| {
-        Err(harness_core::error::HarnessError::AgentExecution(
-            "turn execution ended without agent result".to_string(),
-        ))
-    }) {
-        Ok(()) => match server.thread_manager.complete_turn(&thread_id, &turn_id) {
-            Ok(Some(usage)) => {
-                persist_runtime_thread(&thread_db, &server, &thread_id).await;
-                emit_runtime_notification(
-                    &notify_tx,
-                    &notification_tx,
-                    Notification::TurnCompleted {
-                        turn_id: turn_id.clone(),
-                        status: TurnStatus::Completed,
-                        token_usage: usage,
-                    },
-                );
-            }
-            Ok(None) => {}
-            Err(err) => {
-                let error_msg = err.to_string();
-                tracing::error!(
-                    thread_id = %thread_id,
-                    turn_id = %turn_id,
-                    "failed to complete turn after execution: {error_msg}"
-                );
-                if let Err(e) = server.thread_manager.add_item(
-                    &thread_id,
-                    &turn_id,
-                    harness_core::types::Item::Error {
-                        code: -1,
-                        message: format!("Failed to complete turn: {error_msg}"),
-                    },
-                ) {
-                    tracing::warn!("failed to add error item to turn: {e}");
-                } else {
-                    persist_runtime_thread(&thread_db, &server, &thread_id).await;
-                }
-                mark_turn_failed(
-                    &server,
-                    &thread_db,
-                    &notify_tx,
-                    &notification_tx,
-                    &thread_id,
-                    &turn_id,
-                    error_msg,
-                )
-                .await;
-            }
-        },
-        Err(err) => {
-            let error_msg = err.to_string();
-            if let Err(e) = server.thread_manager.add_item(
-                &thread_id,
-                &turn_id,
-                harness_core::types::Item::Error {
-                    code: -1,
-                    message: error_msg.clone(),
-                },
-            ) {
-                tracing::warn!("failed to add error item to turn: {e}");
-            } else {
-                persist_runtime_thread(&thread_db, &server, &thread_id).await;
-            }
-            mark_turn_failed(
-                &server,
-                &thread_db,
-                &notify_tx,
-                &notification_tx,
-                &thread_id,
-                &turn_id,
-                error_msg,
-            )
-            .await;
-        }
-    }
-}
-
-/// Compute exponential backoff: `min(base_ms * 2^(attempt-1), max_ms)`.
-///
-/// - Attempt 1: `base_ms`
-/// - Attempt 2: `base_ms * 2`
-/// - Attempt 3: `base_ms * 4`
-/// - …capped at `max_ms`
-fn compute_backoff_ms(base_ms: u64, max_ms: u64, attempt: u32) -> u64 {
-    let shift = attempt.saturating_sub(1).min(63);
-    base_ms.saturating_mul(1u64 << shift).min(max_ms)
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-enum ImplementationOutcome {
-    PlanIssue(String),
-    ParsedPr {
-        pr_url: Option<String>,
-        pr_num: Option<u64>,
-        created_issue_num: Option<u64>,
-    },
-}
-
-fn parse_implementation_outcome(output: &str) -> ImplementationOutcome {
-    if let Some(desc) = prompts::parse_plan_issue(output) {
-        return ImplementationOutcome::PlanIssue(desc);
-    }
-    let pr_url = prompts::parse_pr_url(output);
-    let pr_num = pr_url.as_deref().and_then(prompts::extract_pr_number);
-    let created_issue_num = prompts::parse_created_issue_number(output);
-    ImplementationOutcome::ParsedPr {
-        pr_url,
-        pr_num,
-        created_issue_num,
-    }
-}
-
-/// Persist a completed stream item as a task artifact when it carries content
-/// worth retaining across context loss (shell commands, file edits, tool calls).
-async fn persist_artifact(
-    store: &TaskStore,
-    task_id: &TaskId,
-    turn: u32,
-    item: &harness_core::types::Item,
-) {
-    let (artifact_type, content) = match item {
-        harness_core::types::Item::ShellCommand {
-            command,
-            exit_code,
-            stdout,
-            stderr,
-        } => {
-            let c = serde_json::json!({
-                "command": command,
-                "exit_code": exit_code,
-                "stdout": stdout,
-                "stderr": stderr,
-            })
-            .to_string();
-            ("shell_command", c)
-        }
-        harness_core::types::Item::FileEdit {
-            path,
-            before,
-            after,
-        } => {
-            let c = serde_json::json!({
-                "path": path,
-                "before": before,
-                "after": after,
-            })
-            .to_string();
-            ("file_edit", c)
-        }
-        harness_core::types::Item::ToolCall {
-            name,
-            input,
-            output,
-        } => {
-            let c = serde_json::json!({
-                "name": name,
-                "input": input,
-                "output": output,
-            })
-            .to_string();
-            ("tool_call", c)
-        }
-        _ => return,
-    };
-    store
-        .insert_artifact(task_id, turn, artifact_type, &content)
-        .await;
-}
-
-/// Execute an agent request via [`CodeAgent::execute_stream`], broadcasting
-/// each [`StreamItem`] to the per-task channel in real time, and reconstruct
-/// an [`AgentResponse`] from the collected stream events.
-///
-/// `turn` is stored with each captured artifact so callers can distinguish
-/// implementation turn (1) from later review or retry turns.
-async fn run_agent_streaming(
-    agent: &dyn CodeAgent,
-    req: AgentRequest,
-    task_id: &TaskId,
-    store: &TaskStore,
-    turn: u32,
-) -> harness_core::error::Result<(AgentResponse, Option<u64>)> {
-    let turn_start = Instant::now();
-    let (tx, mut rx) = tokio::sync::mpsc::channel::<StreamItem>(128);
-    let mut exec = std::pin::pin!(agent.execute_stream(req, tx));
-    let mut exec_result: Option<harness_core::error::Result<()>> = None;
-    let mut channel_closed = false;
-    let mut output = String::new();
-    let mut token_usage = TokenUsage::default();
-    let mut first_token_latency_ms: Option<u64> = None;
-    // Tracks the last CREATED_ISSUE= number we successfully wrote to the DB so
-    // we can (a) skip redundant writes and (b) detect agent self-corrections
-    // that replace an earlier sentinel with a later one ("last sentinel wins").
-    // Using Option<u64> rather than a boolean means each distinct value causes
-    // exactly one DB write, and a second CREATED_ISSUE=20 after CREATED_ISSUE=10
-    // correctly overwrites the stored external_id.
-    let mut last_backfilled_issue: Option<u64> = None;
-    // Pre-check whether this task is auto-fix to avoid a cache lookup on every
-    // MessageDelta.  The source field is immutable after creation.
-    let is_auto_fix_task = store
-        .get(task_id)
-        .is_some_and(|s| s.source.as_deref() == Some("auto-fix"));
-
-    loop {
-        tokio::select! {
-            result = &mut exec, if exec_result.is_none() => {
-                exec_result = Some(result);
-            }
-            item = rx.recv(), if !channel_closed => {
-                match item {
-                    Some(item) => {
-                        store.publish_stream_item(task_id, item.clone());
-                        match &item {
-                            StreamItem::MessageDelta { text } => {
-                                if first_token_latency_ms.is_none() {
-                                    first_token_latency_ms =
-                                        Some(turn_start.elapsed().as_millis() as u64);
-                                }
-                                output.push_str(text);
-
-                                // Early backfill: scan the full accumulated output on
-                                // every delta.  Scanning the full buffer (rather than
-                                // only the newly appended window) is required for two
-                                // correctness properties:
-                                //
-                                // 1. Chunked streaming: if the sentinel was split across
-                                //    chunk boundaries (e.g. "CREATED_ISSUE=" in one
-                                //    delta, "42\n" in the next), a window-only scan
-                                //    would miss the number.  The full-buffer scan sees
-                                //    the complete sentinel once both chunks have arrived.
-                                //
-                                // 2. "Last sentinel wins": parse_created_issue_number
-                                //    returns the LAST CREATED_ISSUE= in the output, so
-                                //    a later self-correction (CREATED_ISSUE=20 after
-                                //    CREATED_ISSUE=10) is written to the DB.
-                                //    last_backfilled_issue prevents redundant writes
-                                //    when the parsed number has not changed.
-                                if is_auto_fix_task {
-                                    // Only scan up to the last newline so that a
-                                    // sentinel split across chunk boundaries (e.g.
-                                    // "CREATED_ISSUE=4" then "2\n" in the next delta)
-                                    // does not produce a spurious partial value.
-                                    let complete_len =
-                                        output.rfind('\n').map(|i| i + 1).unwrap_or(0);
-                                    if let Some(issue_num) =
-                                        prompts::parse_created_issue_number(
-                                            &output[..complete_len],
-                                        )
-                                    {
-                                        if Some(issue_num) != last_backfilled_issue {
-                                            let eid = format!("issue:{issue_num}");
-                                            match store
-                                                .overwrite_external_id_auto_fix(task_id, &eid)
-                                                .await
-                                            {
-                                                Ok(()) => {
-                                                    tracing::info!(
-                                                        task_id = %task_id,
-                                                        external_id = %eid,
-                                                        "streaming: backfilled external_id for auto-fix task"
-                                                    );
-                                                    last_backfilled_issue = Some(issue_num);
-                                                }
-                                                Err(e) => tracing::warn!(
-                                                    task_id = %task_id,
-                                                    "streaming: failed to backfill external_id: {e}"
-                                                ),
-                                            }
-                                        }
-                                    }
-                                }
-                            }
-                            StreamItem::ItemCompleted {
-                                item: Item::AgentReasoning { content },
-                            } => {
-                                // Non-streaming adapters (e.g. AnthropicApiAgent) call
-                                // execute() to completion before emitting this item, so
-                                // elapsed time here is full-request latency, not TTFB.
-                                // Recording it as first_token_latency_ms would silently
-                                // mix whole-request durations with real streaming TTFB
-                                // values and inflate p50.  Only MessageDelta (true
-                                // streaming events) sets first_token_latency_ms.
-                                // Prefer the full content over accumulated deltas.
-                                output = content.clone();
-                                // For non-streaming adapters the full output arrives
-                                // here; apply the same sentinel scan so backfill
-                                // happens before the post-execution path runs (closes
-                                // the webhook race for these adapters too).
-                                if is_auto_fix_task {
-                                    if let Some(issue_num) =
-                                        prompts::parse_created_issue_number(&output)
-                                    {
-                                        if Some(issue_num) != last_backfilled_issue {
-                                            let eid = format!("issue:{issue_num}");
-                                            match store
-                                                .overwrite_external_id_auto_fix(task_id, &eid)
-                                                .await
-                                            {
-                                                Ok(()) => {
-                                                    tracing::info!(
-                                                        task_id = %task_id,
-                                                        external_id = %eid,
-                                                        "streaming: backfilled external_id for auto-fix task"
-                                                    );
-                                                    last_backfilled_issue = Some(issue_num);
-                                                }
-                                                Err(e) => tracing::warn!(
-                                                    task_id = %task_id,
-                                                    "streaming: failed to backfill external_id: {e}"
-                                                ),
-                                            }
-                                        }
-                                    }
-                                }
-                            }
-                            StreamItem::ItemCompleted { item: completed_item } => {
-                                persist_artifact(store, task_id, turn, completed_item).await;
-                            }
-                            StreamItem::TokenUsage { usage } => {
-                                token_usage = usage.clone();
-                            }
-                            StreamItem::Done => {
-                                channel_closed = true;
-                            }
-                            _ => {}
-                        }
-                    }
-                    None => {
-                        channel_closed = true;
-                    }
-                }
-            }
-        }
-        if exec_result.is_some() && channel_closed {
-            break;
-        }
-    }
-
-    match exec_result.unwrap_or_else(|| {
-        Err(HarnessError::AgentExecution(
-            "agent execution completed without result".into(),
-        ))
-    }) {
-        Ok(()) => Ok((
-            AgentResponse {
-                output,
-                stderr: String::new(),
-                items: Vec::new(),
-                token_usage,
-                model: String::new(),
-                exit_code: Some(0),
-            },
-            first_token_latency_ms,
-        )),
-        Err(e) => Err(e),
-    }
-}
-
-fn prepend_constitution(prompt: String, enabled: bool) -> String {
-    const CONSTITUTION: &str = include_str!("../../../config/constitution.md");
-    if enabled {
-        format!("{CONSTITUTION}\n\n{prompt}")
-    } else {
-        prompt
-    }
-}
-
-/// Run a plan step for a complex prompt-only task when the planning gate has
-/// forced `TaskPhase::Plan`.
-///
-/// This gives the planning gate real effect: instead of silently skipping
-/// to Implement, the agent produces a plan first, which is then threaded into
-/// the implementation prompt just like issue-based plans.
-async fn run_plan_for_prompt(
-    agent: &dyn CodeAgent,
-    store: &TaskStore,
-    task_id: &TaskId,
-    cargo_env: &HashMap<String, String>,
-    project: &Path,
-    req: &CreateTaskRequest,
-) -> anyhow::Result<(Option<String>, prompts::TriageComplexity, u32)> {
-    use crate::task_runner::TaskPhase;
-
-    let prompt_text = req.prompt.as_deref().unwrap_or_default();
-    tracing::info!(task_id = %task_id, "pipeline: starting plan phase for prompt-only task");
-
-    mutate_and_persist(store, task_id, |state| {
-        state.phase = TaskPhase::Plan;
-    })
-    .await?;
-
-    let plan_prompt = format!(
-        "Before implementing, produce a concise implementation plan for the following task.\n\
-         List the files to change, the approach, and any non-obvious design decisions.\n\
-         Do NOT write code yet — planning only.\n\n\
-         Task:\n{prompt_text}"
-    );
-
-    let plan_req = AgentRequest {
-        prompt: plan_prompt,
-        project_root: project.to_path_buf(),
-        env_vars: cargo_env.clone(),
-        execution_phase: Some(ExecutionPhase::Planning),
-        ..Default::default()
-    };
-
-    let turn_timeout = crate::task_runner::effective_turn_timeout(req.turn_timeout_secs);
-    // Use agent.execute() directly — NOT run_agent_streaming — to avoid the
-    // state side-effects that run_agent_streaming performs: PR_URL extraction,
-    // turn counter mutations, and status updates.  Those side-effects would
-    // corrupt the subsequent implement phase (e.g. consuming call #0 from a
-    // mock/real agent that sets pr_url before the real implementation turn).
-    let plan_resp = tokio::time::timeout(turn_timeout, agent.execute(plan_req))
-        .await
-        .map_err(|_| anyhow::anyhow!("plan phase timed out after {}s", req.turn_timeout_secs))?
-        .map_err(|e| anyhow::anyhow!("plan phase agent error: {e}"))?;
-
-    let plan_text = plan_resp.output.clone();
-    mutate_and_persist(store, task_id, |state| {
-        state.plan_output = Some(plan_text.clone());
-        state.phase = TaskPhase::Implement;
-    })
-    .await?;
-    // NOTE: intentionally no write_checkpoint() here.  Prompt-only tasks cannot
-    // be recovered after a crash (http.rs startup recovery hard-fails them because
-    // the original prompt is not persisted — by design, to avoid writing
-    // credentials/customer data to disk).  Writing the plan blob to the checkpoint
-    // table would therefore provide no resumability benefit while writing
-    // prompt-derived content (which often echoes or summarises the raw prompt) to
-    // the on-disk SQLite store, violating the same privacy invariant.
-
-    tracing::info!(task_id = %task_id, plan_len = plan_text.len(), "plan phase complete (prompt-only)");
-    Ok((Some(plan_text), prompts::TriageComplexity::Medium, 1))
-}
-
-/// Run triage → plan pipeline for a fresh issue-based task.
-///
-/// Returns `Some(plan_text)` if the triage decided a plan is needed and the plan
-/// phase completed. Returns `None` only when triage says PROCEED (trivial issue,
-/// skip planning). All failures propagate as errors — no silent fallbacks.
-async fn run_triage_plan_pipeline(
-    agent: &dyn CodeAgent,
-    store: &TaskStore,
-    task_id: &TaskId,
-    issue: u64,
-    cargo_env: &HashMap<String, String>,
-    project: &Path,
-    req: &CreateTaskRequest,
-) -> anyhow::Result<(Option<String>, prompts::TriageComplexity, u32)> {
-    use crate::task_runner::TaskPhase;
-
-    // --- Phase 1: Triage ---
-    tracing::info!(task_id = %task_id, issue, "pipeline: starting triage phase");
-    mutate_and_persist(store, task_id, |state| {
-        state.phase = TaskPhase::Triage;
-    })
-    .await?;
-
-    let triage_prompt = prompts::triage_prompt(issue).to_prompt_string();
-    let triage_req = AgentRequest {
-        prompt: triage_prompt,
-        project_root: project.to_path_buf(),
-        env_vars: cargo_env.clone(),
-        execution_phase: Some(ExecutionPhase::Planning),
-        ..Default::default()
-    };
-
-    let turn_timeout = crate::task_runner::effective_turn_timeout(req.turn_timeout_secs);
-    let (triage_resp, _) = tokio::time::timeout(
-        turn_timeout,
-        run_agent_streaming(agent, triage_req, task_id, store, 0),
-    )
-    .await
-    .map_err(|_| anyhow::anyhow!("triage phase timed out after {}s", req.turn_timeout_secs))?
-    .map_err(|e| anyhow::anyhow!("triage phase agent error: {e}"))?;
-
-    let triage_text = triage_resp.output.clone();
-    mutate_and_persist(store, task_id, |state| {
-        state.triage_output = Some(triage_text.clone());
-    })
-    .await?;
-    if let Err(e) = store
-        .write_checkpoint(task_id, Some(&triage_text), None, None, "triage_done")
-        .await
-    {
-        tracing::warn!(task_id = %task_id, "failed to write triage checkpoint: {e}");
-    }
-
-    let decision = prompts::parse_triage(&triage_resp.output).ok_or_else(|| {
-        anyhow::anyhow!("triage output unparseable — agent did not produce TRIAGE=<decision>")
-    })?;
-    let complexity = prompts::parse_complexity(&triage_resp.output);
-    tracing::info!(task_id = %task_id, ?decision, ?complexity, "triage decision");
-
-    match decision {
-        prompts::TriageDecision::Skip => {
-            mutate_and_persist(store, task_id, |state| {
-                state.status = crate::task_runner::TaskStatus::Done;
-                state.phase = TaskPhase::Terminal;
-                state.error = Some("Triage: skipped — not worth implementing".to_string());
-            })
-            .await?;
-            anyhow::bail!("triage decided to skip issue #{issue}");
-        }
-        prompts::TriageDecision::NeedsClarification => {
-            // Treat as ProceedWithPlan — let the planner figure out ambiguities
-            // instead of failing the task outright.
-            tracing::info!(task_id = %task_id, "triage: NEEDS_CLARIFICATION → treating as PROCEED_WITH_PLAN");
-        }
-        prompts::TriageDecision::Proceed => {
-            tracing::info!(task_id = %task_id, "triage: PROCEED — skipping plan phase");
-            return Ok((None, complexity, 1));
-        }
-        prompts::TriageDecision::ProceedWithPlan => {
-            // Fall through to plan phase.
-        }
-    }
-
-    // --- Phase 2: Plan ---
-    tracing::info!(task_id = %task_id, issue, "pipeline: starting plan phase");
-    mutate_and_persist(store, task_id, |state| {
-        state.phase = TaskPhase::Plan;
-    })
-    .await?;
-
-    let plan_prompt = prompts::plan_prompt(issue, &triage_resp.output).to_prompt_string();
-    let plan_req = AgentRequest {
-        prompt: plan_prompt,
-        project_root: project.to_path_buf(),
-        env_vars: cargo_env.clone(),
-        execution_phase: Some(ExecutionPhase::Planning),
-        ..Default::default()
-    };
-
-    let (plan_resp, _) = tokio::time::timeout(
-        turn_timeout,
-        run_agent_streaming(agent, plan_req, task_id, store, 0),
-    )
-    .await
-    .map_err(|_| anyhow::anyhow!("plan phase timed out after {}s", req.turn_timeout_secs))?
-    .map_err(|e| anyhow::anyhow!("plan phase agent error: {e}"))?;
-
-    let plan_text = plan_resp.output.clone();
-    mutate_and_persist(store, task_id, |state| {
-        state.plan_output = Some(plan_text.clone());
-        state.phase = TaskPhase::Implement;
-    })
-    .await?;
-    if let Err(e) = store
-        .write_checkpoint(task_id, None, Some(&plan_text), None, "plan_done")
-        .await
-    {
-        tracing::warn!(task_id = %task_id, "failed to write plan checkpoint: {e}");
-    }
-
-    tracing::info!(task_id = %task_id, plan_len = plan_text.len(), "plan phase complete");
-    Ok((Some(plan_text), complexity, 2))
-}
-
-/// Fire a single agent turn to rebase a conflicting PR onto `origin/main`.
-///
-/// Returns `true` if the agent reported `REBASE_OK` (i.e. a new commit was
-/// force-pushed), `false` in every other case.  Errors are not propagated —
-/// a rebase failure is not fatal; the review loop will handle the PR.
-async fn run_rebase_turn(
-    agent: &dyn CodeAgent,
-    pr_num: u64,
-    project: &std::path::Path,
-    repo: &str,
-    turn_timeout: Duration,
-    cargo_env: &HashMap<String, String>,
-) -> bool {
-    // Fetch the branch name from GitHub so the prompt has an exact ref.
-    let branch = {
-        let out = tokio::process::Command::new("gh")
-            .current_dir(project)
-            .args([
-                "pr",
-                "view",
-                &pr_num.to_string(),
-                "--json",
-                "headRefName",
-                "--jq",
-                ".headRefName",
-            ])
-            .stdin(std::process::Stdio::null())
-            .stdout(std::process::Stdio::piped())
-            .stderr(std::process::Stdio::null())
-            .output()
-            .await;
-        match out {
-            Ok(o) if o.status.success() => String::from_utf8_lossy(&o.stdout)
-                .trim()
-                .trim_matches('"')
-                .to_string(),
-            _ => {
-                tracing::warn!(
-                    pr = pr_num,
-                    "run_rebase_turn: could not fetch branch name; skipping"
-                );
-                return false;
-            }
-        }
-    };
-
-    // Reject branch names that contain shell metacharacters.  The branch
-    // name is embedded inside single-quoted shell commands in the agent
-    // prompt; a single-quote in the name would break out of the quoting.
-    if !branch
-        .chars()
-        .all(|c| c.is_alphanumeric() || matches!(c, '/' | '-' | '_' | '.' | '@' | '~' | '+' | ':'))
-    {
-        tracing::warn!(
-            pr = pr_num,
-            branch = %branch,
-            "run_rebase_turn: branch name contains unsafe characters; skipping"
-        );
-        return false;
-    }
-
-    let prompt = harness_core::prompts::rebase_conflicting_pr(pr_num, &branch, repo);
-    let req = AgentRequest {
-        prompt,
-        project_root: project.to_path_buf(),
-        env_vars: cargo_env.clone(),
-        execution_phase: Some(harness_core::types::ExecutionPhase::Planning),
-        ..Default::default()
-    };
-
-    match tokio::time::timeout(turn_timeout, agent.execute(req)).await {
-        Ok(Ok(resp)) => {
-            let last = resp.output.lines().next_back().unwrap_or("").trim();
-            if last.contains("REBASE_OK") {
-                tracing::info!(pr = pr_num, "rebase turn: REBASE_OK");
-                true
-            } else {
-                tracing::warn!(
-                    pr = pr_num,
-                    last_line = last,
-                    "rebase turn: REBASE_FAILED or unexpected output"
-                );
-                false
-            }
-        }
-        Ok(Err(e)) => {
-            tracing::warn!(pr = pr_num, error = %e, "rebase turn: agent error");
-            false
-        }
-        Err(_) => {
-            tracing::warn!(pr = pr_num, "rebase turn: timed out");
-            false
-        }
-    }
-}
-
-pub(crate) async fn run_task(
-    store: &TaskStore,
-    task_id: &TaskId,
-    agent: &dyn CodeAgent,
-    reviewer: Option<&dyn CodeAgent>,
-    skills: Arc<RwLock<harness_skills::store::SkillStore>>,
-    events: Arc<harness_observe::event_store::EventStore>,
-    interceptors: Arc<Vec<Arc<dyn harness_core::interceptor::TurnInterceptor>>>,
-    req: &CreateTaskRequest,
-    project: PathBuf,
-    server_config: &harness_core::config::HarnessConfig,
-    // Accumulated turn count from previous transient-retry attempts.
-    // Ensures the max_turns budget is global across the full task lifecycle,
-    // not reset on each retry (fix for budget-reset-on-retry bug).
-    turns_used_acc: &mut u32,
-) -> anyhow::Result<()> {
-    let task_start = Instant::now();
-
-    if !project.exists() {
-        anyhow::bail!("project_root does not exist: {}", project.display());
-    }
-
-    // Set CARGO_TARGET_DIR to a per-task temp path so parallel agents running
-    // cargo check/test simultaneously do not contend on the same build directory.
-    // A per-project path caused `.cargo-lock` contention and build failures when
-    // two tasks targeted the same project concurrently (issue #488).
-    let task_target = std::env::temp_dir()
-        .join("harness-cargo-targets")
-        .join(task_id.as_str());
-    let cargo_env: HashMap<String, String> = [(
-        "CARGO_TARGET_DIR".to_string(),
-        task_target.display().to_string(),
-    )]
-    .into();
-    // Guard ensures the directory is removed when run_task exits, regardless of
-    // the exit path (success, validation failure, timeout, or review exhaustion).
-    let _task_target_guard = TaskTargetDir(task_target);
-
-    let project_config = load_project_config(&project).with_context(|| {
-        format!(
-            "failed to load project config for task {} at {}",
-            task_id.as_str(),
-            project.display()
-        )
-    })?;
-    let resolved = harness_core::config::resolve::resolve_config(server_config, &project_config);
-    let review_config = &resolved.review;
-    let git = Some(&project_config.git);
-    let repo_slug = detect_repo_slug(&project)
-        .await
-        .unwrap_or_else(|| "{owner}/{repo}".to_string());
-
-    // --- Checkpoint-based resume detection ---
-    // Load checkpoint and task state to determine if we can skip phases.
-    // This is the duplicate-PR prevention gate: if the task already has a PR,
-    // we skip triage/plan/implement and jump directly to agent review.
-    //
-    // Check task store for an existing pr_url first — this survives checkpoint
-    // read failures (e.g. transient SQLite contention) and lets us safely
-    // resume review even when the checkpoint row is temporarily unreadable.
-    let task_pr_url: Option<String> = store.get(task_id).and_then(|t| t.pr_url);
-    let checkpoint = match store.load_checkpoint(task_id).await {
-        Ok(cp) => cp,
-        Err(e) => {
-            if task_pr_url.is_some() {
-                // Task state already records a pr_url — safe to resume review
-                // without the checkpoint; log the failure for observability.
-                tracing::warn!(
-                    task_id = %task_id,
-                    error = %e,
-                    "checkpoint load failed but task already has pr_url; resuming review without checkpoint"
-                );
-                None
-            } else {
-                // No pr_url in task state — fail closed to prevent duplicate PR.
-                return Err(e).with_context(|| {
-                    format!(
-                        "failed to load checkpoint for task {}; aborting to prevent duplicate PR",
-                        task_id
-                    )
-                });
-            }
-        }
-    };
-    let resumed_pr_url: Option<String> =
-        task_pr_url.or_else(|| checkpoint.as_ref().and_then(|c| c.pr_url.clone()));
-    // Capture before `resumed_pr_url` is moved into the `'implement` block.
-    let was_resumed_pr = resumed_pr_url.is_some();
-    let resumed_plan: Option<String> = checkpoint.and_then(|c| c.plan_output);
-
-    // --- Pipeline: Triage → Plan → Implement ---
-    // For issue-based tasks without an existing PR, run triage first.
-    // Triage decides whether to skip planning or go through a plan phase.
-    // Checkpoint overrides: if a plan was saved, skip the pipeline entirely.
-    let (plan_output, triage_complexity, pipeline_turns) = if resumed_pr_url.is_some() {
-        // PR already exists — skip triage/plan entirely.
-        (None, prompts::TriageComplexity::Medium, 0u32)
-    } else if let Some(plan) = resumed_plan {
-        // Plan checkpoint found — use saved plan, skip triage/plan pipeline.
-        tracing::info!(task_id = %task_id, "checkpoint resume: using saved plan, skipping triage/plan");
-        (Some(plan), prompts::TriageComplexity::Medium, 0u32)
-    } else if let Some(issue) = req.issue {
-        // Only triage fresh issues (no existing PR to continue).
-        let has_existing_pr = find_existing_pr_for_issue(&project, issue)
-            .await
-            .ok()
-            .flatten()
-            .is_some();
-        if !has_existing_pr {
-            run_triage_plan_pipeline(agent, store, task_id, issue, &cargo_env, &project, req)
-                .await?
-        } else {
-            (None, prompts::TriageComplexity::Medium, 0u32)
-        }
-    } else {
-        // Planning gate (task_runner) may have forced TaskPhase::Plan for a
-        // complex prompt-only task.  Check the stored phase so the gate has
-        // real effect rather than silently falling through to Implement.
-        let forced_plan = store
-            .get(task_id)
-            .map(|s| s.phase == crate::task_runner::TaskPhase::Plan)
-            .unwrap_or(false);
-        if forced_plan && req.issue.is_none() && req.pr.is_none() {
-            // Set to Implementing BEFORE the plan phase so that a crash during planning
-            // leaves the task in 'implementing' status. The startup recovery code will catch
-            // it and fail-close it (no pr_url → mark failed), rather than leaving it stuck
-            // as a plain 'pending' that is never re-dispatched.
-            update_status(store, task_id, TaskStatus::Implementing, 1).await?;
-            run_plan_for_prompt(agent, store, task_id, &cargo_env, &project, req).await?
-        } else {
-            (None, prompts::TriageComplexity::Medium, 0u32)
-        }
-    };
-
-    // Derive dynamic parameters from triage complexity.
-    // Triage provides a DEFAULT only — caller's explicit max_rounds always wins (Fix #2).
-    // Low complexity no longer skips agent review to preserve the review gate (Fix #1).
-    let (triage_default_rounds, skip_agent_review) = match triage_complexity {
-        prompts::TriageComplexity::Low => (2u32, false),
-        prompts::TriageComplexity::Medium => (8u32, false),
-        prompts::TriageComplexity::High => (8u32, false),
-    };
-    let effective_max_rounds = req.max_rounds.unwrap_or(triage_default_rounds);
-    // max_turns: per-request override wins; global config is the fallback.
-    // Counts every agent API call (impl + validation retries + review rounds).
-    let effective_max_turns: Option<u32> = req.max_turns.or(server_config.concurrency.max_turns);
-    // Start from accumulated turns (prior transient-retry attempts + pipeline phases)
-    // so the budget is global across the full task lifecycle.
-    let mut turns_used: u32 = *turns_used_acc + pipeline_turns;
-    *turns_used_acc = turns_used;
-    let jaccard_threshold = server_config.concurrency.loop_jaccard_threshold;
-    tracing::info!(
-        task_id = %task_id,
-        ?triage_complexity,
-        effective_max_rounds,
-        skip_agent_review,
-        ?effective_max_turns,
-        "triage complexity applied"
-    );
-
-    // Resume normal flow — update status to Implementing for the main turn.
-    update_status(store, task_id, TaskStatus::Implementing, 1).await?;
-
-    let first_prompt = if let Some(issue) = req.issue {
-        let base = match find_existing_pr_for_issue(&project, issue).await {
-            Ok(Some((pr_num, branch, pr_url))) => {
-                tracing::info!(
-                    "reusing existing PR #{pr_num} on branch `{branch}` for issue #{issue}"
-                );
-                let slug = prompts::repo_slug_from_pr_url(Some(&pr_url));
-                prompts::continue_existing_pr(issue, pr_num, &branch, &slug)
-            }
-            Ok(None) => {
-                prompts::implement_from_issue(issue, git, plan_output.as_deref()).to_prompt_string()
-            }
-            Err(e) => {
-                tracing::warn!("failed to check for existing PR for issue #{issue}: {e}");
-                prompts::implement_from_issue(issue, git, plan_output.as_deref()).to_prompt_string()
-            }
-        };
-        // If the caller also supplied a description alongside the issue number, include it
-        // as additional context. Without this, batch tasks that set both `description` and
-        // `issue` would silently discard the description.
-        if let Some(hint) = req.prompt.as_deref().filter(|s| !s.is_empty()) {
-            format!(
-                "{base}\n\nAdditional context from caller:\n{}",
-                prompts::wrap_external_data(hint)
-            )
-        } else {
-            base
-        }
-    } else if let Some(pr) = req.pr {
-        prompts::check_existing_pr(
-            pr,
-            &review_config.review_bot_command,
-            &repo_slug,
-            &review_config.reviewer_name,
-            false,
-        )
-    } else if matches!(
-        req.source.as_deref(),
-        Some("periodic_review") | Some("sprint_planner")
-    ) {
-        // Review/planner tasks use their prompt as-is — no "create PR" wrapper.
-        req.prompt.clone().unwrap_or_default()
-    } else {
-        prompts::implement_from_prompt(req.prompt.as_deref().unwrap_or_default(), git)
-    };
-
-    // Inject language-detected validation instructions into the prompt when no
-    // explicit validation config is set. This delegates validation to the agent
-    // instead of running external commands that may not be installed.
-    let first_prompt = if project_config.validation.pre_commit.is_empty()
-        && project_config.validation.pre_push.is_empty()
-    {
-        let lang = lang_detect::detect_language(&project);
-        let instructions = lang_detect::validation_prompt_instructions(lang, &project);
-        if instructions.is_empty() {
-            first_prompt
-        } else {
-            format!("{first_prompt}\n\n{instructions}")
-        }
-    } else {
-        first_prompt
-    };
-
-    // Inject sibling-awareness context when other agents are working on the same project.
-    // This prevents parallel agents from over-scoping their changes into each other's files.
-    //
-    // `project` may be a per-task worktree path when workspace isolation is active, but the
-    // sibling cache stores the canonical source repo path (set at spawn time before the worktree
-    // is created). Look up the canonical root from the task's own cache entry so that
-    // list_siblings() path comparison works correctly in the isolated-worktree case.
-    let first_prompt = {
-        let canonical_project = store
-            .get(task_id)
-            .and_then(|s| s.project_root)
-            .unwrap_or_else(|| project.clone());
-        let siblings = store.list_siblings(&canonical_project, task_id);
-        if siblings.is_empty() {
-            first_prompt
-        } else {
-            let sibling_tasks: Vec<prompts::SiblingTask> = siblings
-                .into_iter()
-                .filter_map(|s| {
-                    s.description.and_then(|description| {
-                        if description.is_empty() {
-                            None
-                        } else {
-                            Some(prompts::SiblingTask {
-                                issue: s.issue,
-                                description,
-                            })
-                        }
-                    })
-                })
-                .collect();
-            let ctx = prompts::sibling_task_context(&sibling_tasks);
-            format!("{first_prompt}\n\n{ctx}")
-        }
-    };
-
-    // Prepend the Golden Principles constitution when enabled.
-    let first_prompt =
-        prepend_constitution(first_prompt, server_config.server.constitution_enabled);
-
-    // Inject skill content directly into the prompt text.
-    // Since harness uses single-turn `claude -p`, context items are not visible
-    // to the agent — we must embed skill content in the prompt string itself.
-    // Also records usage for any matched skills via record_use().
-    let matched_skills = matched_skills_for_prompt(&skills, &first_prompt).await;
-    let skill_additions = inject_skills_into_prompt(&skills, &first_prompt).await;
-    let first_prompt = if skill_additions.is_empty() {
-        first_prompt
-    } else {
-        first_prompt + &skill_additions
-    };
-    for (skill_id, skill_name) in matched_skills {
-        let mut skill_event = Event::new(
-            SessionId::new(),
-            "skill_used",
-            "task_runner",
-            Decision::Pass,
-        );
-        skill_event.reason = Some(skill_name);
-        skill_event.detail = Some(format!(
-            "task_id={} skill_id={}",
-            task_id.as_str(),
-            skill_id.as_str()
-        ));
-        if let Err(err) = events.log(&skill_event).await {
-            tracing::warn!(error = %err, "failed to log skill_used event");
-        }
-    }
-
-    let context_items = collect_context_items(&skills, &project, &first_prompt).await;
-
-    let turn_timeout = crate::task_runner::effective_turn_timeout(req.turn_timeout_secs);
-
-    // Periodic review tasks need Bash to run guard check commands but should
-    // not have unrestricted write access — use Standard profile. All other
-    // tasks (implementation) keep Full (None → --dangerously-skip-permissions).
-    //
-    // Some(tools) causes claude.rs to pass --allowedTools to the CLI (hard
-    // enforcement). None → --dangerously-skip-permissions (full access).
-    let (initial_allowed_tools, capability_prompt_note): (Option<Vec<String>>, _) = if matches!(
-        req.source.as_deref(),
-        Some("periodic_review") | Some("sprint_planner")
-    ) {
-        (
-            Some(restricted_tools(CapabilityProfile::Standard)?),
-            CapabilityProfile::Standard.prompt_note(),
-        )
-    } else {
-        (None, None)
-    };
-
-    // Prepend capability restriction note so the agent knows which tools are
-    // permitted. This is the primary enforcement path now that --allowedTools
-    // is not passed to the CLI.
-    let first_prompt = if let Some(note) = capability_prompt_note {
-        format!("{note}\n\n{first_prompt}")
-    } else {
-        first_prompt
-    };
-
-    tracing::info!(
-        task_id = %task_id,
-        prompt_len = first_prompt.len(),
-        prompt_empty = first_prompt.is_empty(),
-        source = ?req.source,
-        "run_task: prompt constructed"
-    );
-    if first_prompt.is_empty() {
-        tracing::error!(task_id = %task_id, "run_task: prompt is empty — agent will fail");
-    }
-
-    // Duplicate-PR prevention guard: if checkpoint shows PR already created, skip the
-    // implement agent entirely to avoid opening a second PR on resume.
-    let (pr_url, pr_num): (Option<String>, u64) = 'implement: {
-        if let Some(pr_str) = resumed_pr_url {
-            tracing::info!(
-                task_id = %task_id,
-                pr_url = %pr_str,
-                "checkpoint resume: PR exists, skipping implement phase"
-            );
-            let Some(n) = prompts::extract_pr_number(&pr_str) else {
-                tracing::error!(
-                    task_id = %task_id,
-                    pr_url = %pr_str,
-                    "checkpoint resume: cannot parse PR number, marking failed"
-                );
-                mutate_and_persist(store, task_id, |s| {
-                    s.status = TaskStatus::Failed;
-                    s.turn = 1;
-                    s.error = Some(format!("checkpoint resume: no PR number in {pr_str}"));
-                })
-                .await?;
-                return Ok(());
-            };
-            mutate_and_persist(store, task_id, |s| {
-                s.pr_url = Some(pr_str.clone());
-                s.rounds.push(RoundResult {
-                    turn: 1,
-                    action: "implement".into(),
-                    result: "resumed_checkpoint".into(),
-                    detail: None,
-                    first_token_latency_ms: None,
-                });
-            })
-            .await?;
-            break 'implement (Some(pr_str), n);
-        }
-
-        let impl_phase_start = Instant::now();
-
-        let initial_req = AgentRequest {
-            prompt: first_prompt,
-            project_root: project.clone(),
-            context: context_items.clone(),
-            max_budget_usd: req.max_budget_usd,
-            execution_phase: Some(ExecutionPhase::Planning),
-            allowed_tools: initial_allowed_tools,
-            env_vars: cargo_env.clone(),
-            ..Default::default()
-        };
-
-        // Run pre_execute interceptors; Block aborts the task.
-        let first_req = run_pre_execute(&interceptors, initial_req).await?;
-
-        // Execute implementation turn with post-execution validation and auto-retry.
-        // Use the largest max_retries declared by any interceptor.
-        // A single interceptor returning 0 should not suppress retries for others.
-        let max_validation_retries: u32 = interceptors
-            .iter()
-            .filter_map(|i| i.max_validation_retries())
-            .max()
-            .unwrap_or(2);
-        let mut validation_attempt = 0u32;
-        let mut impl_req = first_req.clone();
-
-        let resp = loop {
-            if let Some(max) = effective_max_turns {
-                if turns_used >= max {
-                    return Err(anyhow::anyhow!(
-                        "Turn budget exhausted: used {} of {} allowed turns",
-                        turns_used,
-                        max
-                    ));
-                }
-            }
-            let raw = tokio::time::timeout(
-                turn_timeout,
-                run_agent_streaming(agent, impl_req.clone(), task_id, store, 1),
-            )
-            .await;
-            turns_used += 1;
-            *turns_used_acc = turns_used;
-            match raw {
-                Ok(Ok((r, impl_first_token_ms))) => {
-                    // Post-execution tool isolation check (defense-in-depth alongside
-                    // --allowedTools CLI enforcement). Violations feed into the retry loop
-                    // so the agent gets a chance to self-correct (fail-closed, not fail-open).
-                    let impl_tools = impl_req.allowed_tools.as_deref().unwrap_or(&[]);
-                    let tool_violations = validate_tool_usage(&r.output, impl_tools);
-                    let violation_err: Option<String> = if !tool_violations.is_empty() {
-                        let msg = format!(
-                        "[VALIDATION ERROR] Tool isolation violation: agent used disallowed tools: [{}]. Only [{}] are permitted.",
-                        tool_violations.join(", "),
-                        impl_tools.join(", ")
-                    );
-                        tracing::warn!(
-                            ?tool_violations,
-                            "implementation turn: agent used tools outside allowed list"
-                        );
-                        Some(msg)
-                    } else {
-                        None
-                    };
-                    // PreToolUse / PostToolUse hook injection point:
-                    // detect files written during this turn and fire post_tool_use hooks.
-                    let hook_err = {
-                        let modified = detect_modified_files(&project).await;
-                        if modified.is_empty() {
-                            None
-                        } else {
-                            let hook_event = ToolUseEvent {
-                                tool_name: "file_write".to_string(),
-                                affected_files: modified,
-                                session_id: None,
-                            };
-                            run_post_tool_use(&interceptors, &hook_event, &project).await
-                        }
-                    };
-                    let post_err = run_post_execute(&interceptors, &impl_req, &r).await;
-                    let combined_err = violation_err.or(hook_err).or(post_err);
-                    if let Some(err) = combined_err {
-                        if validation_attempt < max_validation_retries {
-                            validation_attempt += 1;
-                            let backoff_ms = compute_backoff_ms(
-                                req.retry_base_backoff_ms,
-                                req.retry_max_backoff_ms,
-                                validation_attempt,
-                            );
-                            tracing::warn!(
-                                attempt = validation_attempt,
-                                max = max_validation_retries,
-                                backoff_ms,
-                                error = %err,
-                                "post-execution validation failed; backing off before retry"
-                            );
-                            let truncated = truncate_validation_error(&err, 2000);
-                            impl_req.prompt = format!(
-                                "{}\n\nPost-execution validation failed (attempt {}/{}).\nErrors are prefixed with [COMPILE ERROR], [TEST FAILURE], [LINT ERROR], or [VALIDATION ERROR] for classified failures, or with an interceptor name (e.g. [hook_name]) for hook/policy violations — focus your fix on the indicated error type:\n{}",
-                                first_req.prompt,
-                                validation_attempt,
-                                max_validation_retries,
-                                truncated
-                            );
-                            sleep(Duration::from_millis(backoff_ms)).await;
-                            continue;
-                        } else {
-                            tracing::error!(
-                                max = max_validation_retries,
-                                error = %err,
-                                "post-execution validation failed after max retries; aborting task"
-                            );
-                            run_on_error(&interceptors, &impl_req, &err).await;
-                            return Err(anyhow::anyhow!(
-                                "Post-execution validation failed after {} attempts: {}",
-                                max_validation_retries,
-                                err
-                            ));
-                        }
-                    }
-                    break (r, impl_first_token_ms);
-                }
-                Ok(Err(e)) => {
-                    run_on_error(&interceptors, &impl_req, &e.to_string()).await;
-                    return Err(e.into());
-                }
-                Err(_) => {
-                    let msg = format!(
-                        "Implementation turn timed out after {}s",
-                        turn_timeout.as_secs()
-                    );
-                    run_on_error(&interceptors, &impl_req, &msg).await;
-                    return Err(anyhow::anyhow!("{msg}"));
-                }
-            }
-        };
-
-        let (
-            AgentResponse {
-                output,
-                stderr,
-                token_usage: impl_token_usage,
-                ..
-            },
-            impl_first_token_ms,
-        ) = resp;
-
-        tracing::info!(
-            task_id = %task_id,
-            phase = "implementing",
-            elapsed_secs = impl_phase_start.elapsed().as_secs(),
-            "phase_completed"
-        );
-        {
-            let preview: String = output.chars().take(200).collect();
-            tracing::info!(
-                task_id = %task_id,
-                output_chars = output.len(),
-                preview = %preview,
-                input_tokens = impl_token_usage.input_tokens,
-                output_tokens = impl_token_usage.output_tokens,
-                "agent_output_summary"
-            );
-        }
-
-        if !stderr.is_empty() {
-            tracing::warn!(stderr = %stderr, "agent stderr during implementation");
-        }
-
-        // Review-only tasks produce a report, not a PR.
-        // Persist the output and return immediately — no PR parsing or review loop.
-        let is_review_task = store.get(task_id).is_some_and(|s| {
-            matches!(
-                s.source.as_deref(),
-                Some("periodic_review") | Some("sprint_planner")
-            )
-        });
-
-        if is_review_task {
-            mutate_and_persist(store, task_id, |s| {
-                s.status = TaskStatus::Done;
-                s.turn = 1;
-                s.rounds.push(RoundResult {
-                    turn: 1,
-                    action: "review".into(),
-                    result: "completed".into(),
-                    detail: if output.is_empty() {
-                        None
-                    } else {
-                        Some(output.clone())
-                    },
-                    first_token_latency_ms: impl_first_token_ms,
-                });
-            })
-            .await?;
-            store.log_event(crate::event_replay::TaskEvent::Completed {
-                task_id: task_id.0.clone(),
-                ts: crate::event_replay::now_ts(),
-            });
-            tracing::info!(
-                task_id = %task_id,
-                status = "done",
-                turns = 1,
-                pr_url = tracing::field::Empty,
-                total_elapsed_secs = task_start.elapsed().as_secs(),
-                "task_completed"
-            );
-            return Ok(());
-        }
-
-        let (pr_url, pr_num, created_issue_num) = match parse_implementation_outcome(&output) {
-            ImplementationOutcome::PlanIssue(plan_issue) => {
-                tracing::error!(
-                    task_id = %task_id,
-                    plan_issue = %plan_issue,
-                    "implementation returned PLAN_ISSUE; marking task failed"
-                );
-                mutate_and_persist(store, task_id, |s| {
-                    s.status = TaskStatus::Failed;
-                    s.turn = 2;
-                    s.error = Some(plan_issue.clone());
-                    s.rounds.push(RoundResult {
-                        turn: 1,
-                        action: "implement".into(),
-                        result: "plan_issue".into(),
-                        detail: if output.is_empty() {
-                            None
-                        } else {
-                            Some(output.clone())
-                        },
-                        first_token_latency_ms: impl_first_token_ms,
-                    });
-                })
-                .await?;
-                tracing::info!(
-                    task_id = %task_id,
-                    status = "failed",
-                    turns = 2,
-                    pr_url = tracing::field::Empty,
-                    total_elapsed_secs = task_start.elapsed().as_secs(),
-                    "task_completed"
-                );
-                return Ok(());
-            }
-            ImplementationOutcome::ParsedPr {
-                pr_url,
-                pr_num,
-                created_issue_num,
-            } => (pr_url, pr_num, created_issue_num),
-        };
-
-        mutate_and_persist(store, task_id, |s| {
-            s.pr_url = pr_url.clone();
-            s.rounds.push(RoundResult {
-                turn: 1,
-                action: "implement".into(),
-                result: if pr_num.is_some() {
-                    "pr_created".into()
-                } else {
-                    "no_pr".into()
-                },
-                detail: if output.is_empty() {
-                    None
-                } else {
-                    Some(output.clone())
-                },
-                first_token_latency_ms: impl_first_token_ms,
-            });
-        })
-        .await?;
-
-        // Back-fill (or correct) external_id for auto-fix tasks that created a
-        // GitHub issue.  Dedup layers 1-3 match on external_id; without this,
-        // intake re-creates a duplicate task when it receives the webhook for the
-        // newly created issue.
-        //
-        // We compare against the current in-memory value rather than checking
-        // IS NULL so that a post-run self-correction (agent emitted CREATED_ISSUE=20
-        // after an earlier CREATED_ISSUE=10) is always written even when the
-        // streaming path already stored the stale value.
-        if let Some(issue_num) = created_issue_num {
-            let final_eid = format!("issue:{issue_num}");
-            let needs_backfill = store.get(task_id).is_some_and(|s| {
-                s.source.as_deref() == Some("auto-fix")
-                    && s.external_id.as_deref() != Some(&final_eid)
-            });
-            if needs_backfill {
-                if let Err(e) = store
-                    .overwrite_external_id_auto_fix(task_id, &final_eid)
-                    .await
-                {
-                    tracing::warn!(task_id = %task_id, "failed to back-fill external_id: {e}");
-                } else {
-                    tracing::info!(task_id = %task_id, external_id = %final_eid, "back-filled external_id for auto-fix task");
-                }
-            }
-        }
-
-        // Emit PrDetected event so crash recovery can reconstruct pr_url.
-        if let Some(pr_url_str) = pr_url.as_deref() {
-            store.log_event(crate::event_replay::TaskEvent::PrDetected {
-                task_id: task_id.0.clone(),
-                ts: crate::event_replay::now_ts(),
-                pr_url: pr_url_str.to_string(),
-            });
-        }
-
-        // Write PR checkpoint immediately after pr_url is persisted.
-        // This is the most critical checkpoint — it prevents duplicate PR creation on resume.
-        if let Some(pr_url_str) = pr_url.as_deref() {
-            if let Err(e) = store
-                .write_checkpoint(task_id, None, None, Some(pr_url_str), "pr_created")
-                .await
-            {
-                tracing::warn!(task_id = %task_id, "failed to write pr checkpoint: {e}");
-            }
-        }
-
-        // Log implementation event
-        let mut ev = Event::new(
-            SessionId::new(),
-            "task_implement",
-            "task_runner",
-            harness_core::types::Decision::Complete,
-        );
-        ev.detail = pr_num.map(|n| format!("pr={n}"));
-        if let Err(e) = events.log(&ev).await {
-            tracing::warn!("failed to log task_implement event: {e}");
-        }
-
-        let Some(pr_num) = pr_num else {
-            tracing::warn!("no PR number found in agent output; skipping review");
-            mutate_and_persist(store, task_id, |s| {
-                s.status = TaskStatus::Done;
-                s.turn = 2;
-            })
-            .await?;
-            store.log_event(crate::event_replay::TaskEvent::Completed {
-                task_id: task_id.0.clone(),
-                ts: crate::event_replay::now_ts(),
-            });
-            tracing::info!(
-                task_id = %task_id,
-                status = "done",
-                turns = 2,
-                pr_url = tracing::field::Empty,
-                total_elapsed_secs = task_start.elapsed().as_secs(),
-                "task_completed"
-            );
-            return Ok(());
-        };
-
-        (pr_url, pr_num)
-    }; // end 'implement
-
-    // Conflict resolution gate: intercept CONFLICTING PRs before the review loop.
-    // Only runs on the resume/recovery path — fresh PRs cannot be conflicting yet.
-    let mut rebase_pushed = false;
-    if was_resumed_pr {
-        use conflict_resolver::{assess_pr_conflict, PrConflictSize};
-        let conflict_size = assess_pr_conflict(pr_num, &project).await;
-        match conflict_size {
-            PrConflictSize::Small {
-                file_count,
-                region_count,
-            } => {
-                tracing::info!(
-                    pr = pr_num,
-                    file_count,
-                    region_count,
-                    "conflict: small — attempting auto-rebase"
-                );
-                rebase_pushed = run_rebase_turn(
-                    agent,
-                    pr_num,
-                    &project,
-                    &repo_slug,
-                    turn_timeout,
-                    &cargo_env,
-                )
-                .await;
-            }
-            PrConflictSize::Large {
-                file_count,
-                region_count,
-            } => {
-                tracing::warn!(
-                    pr = pr_num,
-                    file_count,
-                    region_count,
-                    "conflict: too large for auto-rebase — deferring to review agent"
-                );
-            }
-            PrConflictSize::Clean | PrConflictSize::Unknown(_) => {}
-        }
-    }
-
-    // Agent review loop (if enabled and reviewer available, and not skipped by triage complexity)
-    let mut agent_pushed_commit = false;
-    if review_config.enabled && !skip_agent_review {
-        if let Some(reviewer) = reviewer {
-            tracing::info!(pr_url = %pr_url.as_deref().unwrap_or(""), "starting agent review");
-            let (review_ok, pushed) = run_agent_review(
-                store,
-                task_id,
-                agent,
-                reviewer,
-                review_config,
-                &context_items,
-                &project,
-                &interceptors,
-                turn_timeout,
-                pr_url.as_deref().unwrap_or(""),
-                project_config.review_type.as_str(),
-                &events,
-                &cargo_env,
-                effective_max_turns,
-                &mut turns_used,
-            )
-            .await?;
-            *turns_used_acc = turns_used;
-            if !review_ok {
-                return Ok(());
-            }
-            agent_pushed_commit = pushed;
-        } else {
-            tracing::warn!("agent review enabled but no reviewer agent configured; skipping");
-        }
-    }
-
-    // Skip external review bot wait when auto-trigger is disabled — there is
-    // no bot to wait for, so the loop would always exhaust all rounds and fail.
-    if !review_config.review_bot_auto_trigger {
-        tracing::info!("review_bot_auto_trigger disabled; skipping external review wait");
-        mutate_and_persist(store, task_id, |s| {
-            s.status = TaskStatus::Done;
-            s.turn = 2;
-        })
-        .await?;
-        store.log_event(crate::event_replay::TaskEvent::Completed {
-            task_id: task_id.0.clone(),
-            ts: crate::event_replay::now_ts(),
-        });
-        tracing::info!(
-            task_id = %task_id,
-            status = "done",
-            turns = 2,
-            pr_url = pr_url.as_deref().unwrap_or(""),
-            total_elapsed_secs = task_start.elapsed().as_secs(),
-            "task_completed"
-        );
-        return Ok(());
-    }
-
-    // Wait for external review bot.
-    // Use a local counter instead of querying the store to derive waiting_count —
-    // task execution is sequential within a single tokio task, so a plain u32 suffices.
-    let mut waiting_count: u32 = 0;
-    waiting_count += 1;
-    update_status(store, task_id, TaskStatus::Waiting, waiting_count).await?;
-
-    let wait_secs = resolved.review_wait_secs.unwrap_or(req.wait_secs);
-    // Project-level override takes precedence over triage-derived rounds so that
-    // per-repo caps (review_max_rounds in harness.toml) are never silently bypassed.
-    let max_rounds = resolved.review_max_rounds.unwrap_or(effective_max_rounds);
-    tracing::info!("waiting {wait_secs}s for review bot on PR #{pr_num}");
-    sleep(Duration::from_secs(wait_secs)).await;
-
-    let review_phase_start = Instant::now();
-
-    // `prev_fixed` tracks whether a commit was pushed that hasn't been re-reviewed yet.
-    // Starts true if the implementation phase pushed a commit, and is set to true again
-    // whenever a review round produces FIXED (agent commits + pushes). This drives the
-    // freshness check in every round after a fix commit, not just round 1.
-    let mut prev_fixed = agent_pushed_commit || rebase_pushed;
-    let repo_slug = prompts::repo_slug_from_pr_url(pr_url.as_deref());
-
-    // Convergence tracking: detect when issue count stops decreasing.
-    let mut issue_counts: Vec<Option<u32>> = Vec::new();
-    let mut impasse = false;
-
-    // Carries test gate failure output from a rejected LGTM into the next review round.
-    let mut pending_test_failure: Option<String> = None;
-    // Issue 2 fix: track whether any LGTM was ever rejected by the test gate
-    // without a subsequent clean LGTM+pass. `pending_test_failure.take()` clears
-    // the Option at the start of each round (to avoid re-showing the same output),
-    // but the graduation check must still know a test gate rejection occurred.
-    let mut lgtm_test_gate_rejected = false;
-
-    // Tracks the most recent non-waiting review output for Jaccard loop detection.
-    let mut prev_review_output: Option<String> = None;
-
-    // Review loop.
-    // Use an explicit counter so WAITING responses don't consume a round — `continue`
-    // inside a `for` loop would silently advance the iterator even without a real review.
-    let mut round: u32 = 1;
-    while round <= max_rounds {
-        update_status(store, task_id, TaskStatus::Reviewing, round).await?;
-
-        let base_prompt = prompts::review_prompt(
-            req.issue,
-            pr_num,
-            round,
-            prev_fixed,
-            &review_config.review_bot_command,
-            &review_config.reviewer_name,
-            &repo_slug,
-            impasse,
-        );
-        // If the previous LGTM was rejected by the test gate, prepend the failure
-        // output so the agent has context for why re-work is needed.
-        let round_prompt = if let Some(failure) = pending_test_failure.take() {
-            format!(
-                "IMPORTANT: The previous LGTM was rejected because the project's tests failed. \
-                 Fix the test failures before declaring LGTM again.\n\n\
-                 Test output:\n```\n{failure}\n```\n\n{base_prompt}"
-            )
-        } else {
-            base_prompt
-        };
-
-        let check_req = AgentRequest {
-            prompt: round_prompt,
-            project_root: project.clone(),
-            context: context_items.clone(),
-            execution_phase: Some(ExecutionPhase::Execution),
-            env_vars: cargo_env.clone(),
-            ..Default::default()
-        };
-        let check_req = run_pre_execute(&interceptors, check_req).await?;
-
-        if let Some(max) = effective_max_turns {
-            if turns_used >= max {
-                let msg = format!(
-                    "Turn budget exhausted: used {} of {} allowed turns",
-                    turns_used, max
-                );
-                tracing::warn!(task_id = %task_id, turns_used, max, "turn budget exhausted in review loop");
-                mutate_and_persist(store, task_id, |s| {
-                    s.status = TaskStatus::Failed;
-                    s.error = Some(msg.clone());
-                })
-                .await?;
-                return Ok(());
-            }
-        }
-        let resp = tokio::time::timeout(turn_timeout, agent.execute(check_req.clone())).await;
-        turns_used += 1;
-        *turns_used_acc = turns_used;
-        let resp = match resp {
-            Ok(Ok(r)) => {
-                let tool_violations = validate_tool_usage(
-                    &r.output,
-                    check_req.allowed_tools.as_deref().unwrap_or(&[]),
-                );
-                if !tool_violations.is_empty() {
-                    let msg = format!(
-                        "Tool isolation violation in review check round {round}: agent used disallowed tools: [{}]",
-                        tool_violations.join(", ")
-                    );
-                    tracing::warn!(
-                        round,
-                        ?tool_violations,
-                        "review check: agent used tools outside allowed list"
-                    );
-                    run_on_error(&interceptors, &check_req, &msg).await;
-                    return Err(anyhow::anyhow!("{msg}"));
-                }
-                if let Some(val_err) = run_post_execute(&interceptors, &check_req, &r).await {
-                    tracing::error!(
-                        round,
-                        error = %val_err,
-                        "post-execute validation failed in review check; will re-enter review"
-                    );
-                    pending_test_failure =
-                        Some(format!("Post-execution validation failed:\n{val_err}"));
-                }
-                r
-            }
-            Ok(Err(e)) => {
-                // Quota/billing failures are not retryable — break immediately instead of
-                // burning remaining review rounds on repeated errors.
-                // Do NOT activate the global rate-limit circuit breaker: the reviewer
-                // agent is configured independently from the implementation agent and a
-                // depleted reviewer account must not stall unrelated implementation tasks.
-                if matches!(
-                    e,
-                    HarnessError::QuotaExhausted(_) | HarnessError::BillingFailed(_)
-                ) {
-                    tracing::error!(round, error = %e, "quota/billing failure during review — aborting review loop");
-                    run_on_error(&interceptors, &check_req, &e.to_string()).await;
-                    mutate_and_persist(store, task_id, |s| {
-                        s.status = TaskStatus::Failed;
-                        s.error = Some(e.to_string());
-                    })
-                    .await?;
-                    return Ok(());
-                }
-                run_on_error(&interceptors, &check_req, &e.to_string()).await;
-                return Err(e.into());
-            }
-            Err(_) => {
-                let msg = format!(
-                    "Review check round {round} timed out after {}s",
-                    turn_timeout.as_secs()
-                );
-                run_on_error(&interceptors, &check_req, &msg).await;
-                return Err(anyhow::anyhow!("{msg}"));
-            }
-        };
-
-        let AgentResponse { output, stderr, .. } = resp;
-
-        if !stderr.is_empty() {
-            tracing::warn!(round, stderr = %stderr, "agent stderr during review check");
-        }
-
-        let raw_lgtm = prompts::is_lgtm(&output);
-        let waiting = prompts::is_waiting(&output);
-        // If post-execute validation failed this round, block LGTM acceptance even
-        // if the reviewer approved — the local validator caught an issue that must be
-        // fixed before the PR can be marked done.
-        let lgtm = raw_lgtm && pending_test_failure.is_none();
-        let fixed = !lgtm && !waiting;
-
-        // Parse issue count before the Jaccard check so loop detection can distinguish
-        // genuine forward progress (decreasing issue count) from true stuck loops.
-        let current_issues = prompts::parse_issue_count(&output);
-
-        // Jaccard loop detection: two consecutive non-waiting, non-LGTM outputs that are
-        // too similar indicate the reviewer is stuck repeating itself without making progress.
-        // Skip when the raw output is an approval: repeated LGTM is legitimate convergence
-        // (e.g., reviewer approves again after a test-gate previously blocked acceptance).
-        if !waiting && !raw_lgtm {
-            if let Some(ref prev) = prev_review_output {
-                let score = jaccard_word_similarity(prev, &output);
-                if score >= jaccard_threshold {
-                    // If the issue count decreased since the previous round, the agent is making
-                    // genuine progress despite similar output structure (e.g., one fix per round
-                    // with a consistent report format). Only abort on true stagnation.
-                    let last_count = issue_counts.last().and_then(|x| *x);
-                    let progressing = matches!(
-                        (last_count, current_issues),
-                        (Some(prev_n), Some(curr_n)) if curr_n < prev_n
-                    );
-                    if !progressing {
-                        let msg = format!(
-                            "review loop detected: output similarity {score:.2} >= threshold {jaccard_threshold:.2}"
-                        );
-                        tracing::warn!(task_id = %task_id, round, score, jaccard_threshold, "jaccard loop detected in review");
-                        mutate_and_persist(store, task_id, |s| {
-                            s.status = TaskStatus::Failed;
-                            s.error = Some(msg.clone());
-                        })
-                        .await?;
-                        return Ok(());
-                    }
-                }
-            }
-            prev_review_output = Some(output.clone());
-        } else if raw_lgtm {
-            // A test-gate-rejected LGTM still resets the comparison baseline so the
-            // next genuine non-LGTM review is not incorrectly compared against a
-            // pre-LGTM review and falsely flagged as a loop.
-            prev_review_output = None;
-        }
-
-        if !waiting {
-            issue_counts.push(current_issues);
-        }
-
-        // Convergence check: if the last 3 rounds all reported issues and the
-        // count is not decreasing, enter impasse mode (critical-only fixes).
-        if issue_counts.len() >= 3 && !impasse {
-            let tail = &issue_counts[issue_counts.len() - 3..];
-            if let [Some(a), Some(b), Some(c)] = tail {
-                if c >= a && c >= b {
-                    impasse = true;
-                    tracing::warn!(
-                        task_id = %task_id,
-                        round,
-                        issues = %format_args!("[{a}, {b}, {c}]"),
-                        "review loop impasse detected: issue count not decreasing"
-                    );
-                }
-            }
-        }
-
-        // WAITING means review bot hasn't posted yet (e.g., quota exhausted).
-        // Don't consume a round — just sleep and retry without incrementing.
-        // Cap consecutive waits to prevent infinite loops when the bot never responds.
-        const MAX_CONSECUTIVE_WAITS: u32 = 10;
-        if waiting {
-            if waiting_count >= MAX_CONSECUTIVE_WAITS {
-                tracing::warn!(
-                    round,
-                    waiting_count,
-                    "PR #{pr_num} review bot has not responded after {MAX_CONSECUTIVE_WAITS} waits; \
-                     consuming a round to prevent infinite loop"
-                );
-                round += 1;
-                waiting_count = 0;
-            } else {
-                tracing::info!(
-                    round,
-                    waiting_count,
-                    "PR #{pr_num} review bot has not responded yet; sleeping without consuming round"
-                );
-                waiting_count += 1;
-                update_status(store, task_id, TaskStatus::Waiting, waiting_count).await?;
-                sleep(Duration::from_secs(wait_secs)).await;
-                continue;
-            }
-        }
-
-        let result_label = if lgtm { "lgtm" } else { "fixed" };
-        mutate_and_persist(store, task_id, |s| {
-            s.rounds.push(RoundResult {
-                turn: round,
-                action: "review".into(),
-                result: result_label.into(),
-                detail: None,
-                first_token_latency_ms: None,
-            });
-        })
-        .await?;
-
-        // Emit RoundCompleted for crash recovery.
-        store.log_event(crate::event_replay::TaskEvent::RoundCompleted {
-            task_id: task_id.0.clone(),
-            ts: crate::event_replay::now_ts(),
-            round,
-            result: result_label.to_string(),
-        });
-
-        // Log pr_review event for observability and GC signal detection.
-        let mut ev = Event::new(
-            SessionId::new(),
-            "pr_review",
-            "task_runner",
-            if lgtm {
-                harness_core::types::Decision::Complete
-            } else {
-                harness_core::types::Decision::Warn
-            },
-        );
-        ev.detail = Some(format!("pr={pr_num}"));
-        let result_label = if lgtm { "lgtm" } else { "fixed" };
-        ev.reason = Some(format!("round {round}: {result_label}"));
-        if let Err(e) = events.log(&ev).await {
-            tracing::warn!("failed to log pr_review event: {e}");
-        }
-
-        if lgtm {
-            // Hard gate: run the project's tests before accepting LGTM.
-            // This prevents agents from gaming the review by manipulating tests
-            // rather than fixing code (OpenAI "Monitoring Reasoning Models", 2026).
-            match run_test_gate(
-                &project,
-                &project_config.validation.pre_push,
-                project_config.validation.test_gate_timeout_secs,
-                &cargo_env,
-            )
-            .await
-            {
-                Ok(()) => {
-                    tracing::info!("PR #{pr_num} approved at round {round}");
-                    mutate_and_persist(store, task_id, |s| {
-                        s.status = TaskStatus::Done;
-                        s.turn = round.saturating_add(1);
-                    })
-                    .await?;
-                    store.log_event(crate::event_replay::TaskEvent::Completed {
-                        task_id: task_id.0.clone(),
-                        ts: crate::event_replay::now_ts(),
-                    });
-                    tracing::info!(
-                        task_id = %task_id,
-                        phase = "reviewing",
-                        elapsed_secs = review_phase_start.elapsed().as_secs(),
-                        "phase_completed"
-                    );
-                    tracing::info!(
-                        task_id = %task_id,
-                        status = "done",
-                        turns = round.saturating_add(1),
-                        pr_url = pr_url.as_deref().unwrap_or(""),
-                        total_elapsed_secs = task_start.elapsed().as_secs(),
-                        "task_completed"
-                    );
-                    return Ok(());
-                }
-                Err(output) => {
-                    tracing::warn!(
-                        task_id = %task_id,
-                        round,
-                        "LGTM rejected: tests failed in test gate, re-entering review round {}",
-                        round.saturating_add(1)
-                    );
-                    lgtm_test_gate_rejected = true;
-                    pending_test_failure = Some(output);
-                }
-            }
-        }
-
-        // Track whether this round produced a fix so the next round enforces
-        // freshness checks against new reviewer feedback.
-        // Also treat a test gate rejection as a "fixed" round — the agent needs
-        // to push a fix and get a fresh review before LGTM can be accepted.
-        prev_fixed = fixed || pending_test_failure.is_some();
-        tracing::info!("PR #{pr_num} fixed at round {round}; waiting for bot re-review");
-        if round < max_rounds {
-            waiting_count += 1;
-            update_status(store, task_id, TaskStatus::Waiting, waiting_count).await?;
-            sleep(Duration::from_secs(wait_secs)).await;
-        }
-        round += 1;
-    }
-
-    // Graduated exit: if the last round had few issues remaining, mark as done
-    // with a warning instead of outright failure — the PR is likely mergeable.
-    // But never graduate when the test gate was still failing on the last round:
-    // a PR whose tests consistently fail must not be approved.
-    let last_issue_count = issue_counts.iter().rev().find_map(|c| *c);
-    // Never graduate when a LGTM was rejected by the test gate — the pending_test_failure
-    // Option was already cleared by take() at round start, so we track rejection separately.
-    let graduated = last_issue_count.is_some_and(|n| n <= 2) && !lgtm_test_gate_rejected;
-
-    if graduated {
-        tracing::info!(
-            task_id = %task_id,
-            remaining_issues = last_issue_count.unwrap_or(0),
-            "review loop exhausted but few issues remain — marking done with warning"
-        );
-        mutate_and_persist(store, task_id, |s| {
-            s.status = TaskStatus::Done;
-            s.turn = max_rounds.saturating_add(1);
-            s.error = Some(format!(
-                "Graduated: LGTM not received after {} rounds but only {} issues remain.",
-                max_rounds,
-                last_issue_count.unwrap_or(0),
-            ));
-        })
-        .await?;
-    } else {
-        mutate_and_persist(store, task_id, |s| {
-            s.status = TaskStatus::Failed;
-            s.turn = max_rounds.saturating_add(1);
-            s.error = Some(format!(
-                "Task did not receive LGTM after {} review rounds (last issues: {}).",
-                max_rounds,
-                last_issue_count.map_or("unknown".to_string(), |n| n.to_string()),
-            ));
-        })
-        .await?;
-    }
-    tracing::info!(
-        task_id = %task_id,
-        phase = "reviewing",
-        elapsed_secs = review_phase_start.elapsed().as_secs(),
-        "phase_completed"
-    );
-    tracing::info!(
-        task_id = %task_id,
-        status = if graduated { "done" } else { "failed" },
-        turns = max_rounds.saturating_add(1),
-        pr_url = pr_url.as_deref().unwrap_or(""),
-        total_elapsed_secs = task_start.elapsed().as_secs(),
-        "task_completed"
-    );
-    Ok(())
-}
-
-/// Compute Jaccard word-similarity between two strings.
-///
-/// Tokenizes each string into a set of non-empty words (split on non-alphanumeric chars),
-/// then returns |intersection| / |union|.
-/// Both empty → 1.0; one empty → 0.0.
-fn jaccard_word_similarity(a: &str, b: &str) -> f64 {
-    let tokens_a: HashSet<&str> = a
-        .split(|c: char| !c.is_alphanumeric())
-        .filter(|s| !s.is_empty())
-        .collect();
-    let tokens_b: HashSet<&str> = b
-        .split(|c: char| !c.is_alphanumeric())
-        .filter(|s| !s.is_empty())
-        .collect();
-
-    if tokens_a.is_empty() && tokens_b.is_empty() {
-        return 1.0;
-    }
-    if tokens_a.is_empty() || tokens_b.is_empty() {
-        return 0.0;
-    }
-
-    let intersection = tokens_a.intersection(&tokens_b).count();
-    let union = tokens_a.union(&tokens_b).count();
-    intersection as f64 / union as f64
-}
-
-/// Normalize a set of review issues into a canonical ordered form.
-/// Issues are sorted by reference before collecting so that insertion order does not affect
-/// equality comparisons, and strings are not unnecessarily cloned during the sort.
-fn normalize_issues(issues: &[String]) -> Vec<String> {
-    let mut sorted: Vec<_> = issues.iter().collect();
-    sorted.sort();
-    sorted.into_iter().cloned().collect()
-}
-
-#[allow(clippy::too_many_arguments)]
-async fn run_agent_review(
-    store: &TaskStore,
-    task_id: &TaskId,
-    agent: &dyn CodeAgent,
-    reviewer: &dyn CodeAgent,
-    review_config: &harness_core::config::agents::AgentReviewConfig,
-    context_items: &[ContextItem],
-    project: &Path,
-    interceptors: &[Arc<dyn harness_core::interceptor::TurnInterceptor>],
-    turn_timeout: Duration,
-    pr_url: &str,
-    project_type: &str,
-    events: &harness_observe::event_store::EventStore,
-    cargo_env: &HashMap<String, String>,
-    effective_max_turns: Option<u32>,
-    turns_used: &mut u32,
-) -> anyhow::Result<(bool, bool)> {
-    let max_rounds = review_config.max_rounds;
-    // (normalized_issues, consecutive_count): tracks how many consecutive rounds produced identical issues.
-    let mut impasse_tracker: Option<(Vec<String>, u32)> = None;
-    // Whether the last action in this loop pushed a new commit (fix or intervention).
-    let mut pushed_commit = false;
-    for agent_round in 1..=max_rounds {
-        update_status(store, task_id, TaskStatus::AgentReview, agent_round).await?;
-
-        // Reviewer evaluates the PR diff — read-only except Bash for `gh pr diff`.
-        let review_req = AgentRequest {
-            prompt: {
-                let base = prompts::agent_review_prompt(pr_url, agent_round, project_type);
-                // Inject capability note — primary enforcement now that --allowedTools
-                // is not passed to the CLI (issue #483).
-                let note = "Tool restriction: you are operating in review mode. \
-                     Only Read, Grep, Glob, and Bash are permitted. \
-                     Use Bash ONLY for read-only commands like `gh pr diff`. \
-                     Do NOT call Write, Edit, or any other tool.";
-                format!("{note}\n\n{base}")
-            },
-            project_root: project.to_path_buf(),
-            context: context_items.to_vec(),
-            execution_phase: Some(ExecutionPhase::SimpleReview),
-            allowed_tools: Some(vec![
-                "Read".to_string(),
-                "Grep".to_string(),
-                "Glob".to_string(),
-                "Bash".to_string(),
-            ]),
-            env_vars: cargo_env.clone(),
-            ..Default::default()
-        };
-        let review_req = run_pre_execute(interceptors, review_req).await?;
-
-        if let Some(max) = effective_max_turns {
-            if *turns_used >= max {
-                return Err(anyhow::anyhow!(
-                    "Turn budget exhausted before agent review round {agent_round}: used {} of {} allowed turns",
-                    turns_used, max
-                ));
-            }
-        }
-        let resp = tokio::time::timeout(turn_timeout, reviewer.execute(review_req.clone())).await;
-        *turns_used += 1;
-        let resp = match resp {
-            Ok(Ok(r)) => {
-                let tool_violations = validate_tool_usage(
-                    &r.output,
-                    review_req.allowed_tools.as_deref().unwrap_or(&[]),
-                );
-                if !tool_violations.is_empty() {
-                    let msg = format!(
-                        "Tool isolation violation in agent review round {agent_round}: agent used disallowed tools: [{}]",
-                        tool_violations.join(", ")
-                    );
-                    tracing::warn!(
-                        agent_round,
-                        ?tool_violations,
-                        "agent review: agent used tools outside allowed list"
-                    );
-                    run_on_error(interceptors, &review_req, &msg).await;
-                    return Err(anyhow::anyhow!("{msg}"));
-                }
-                if let Some(val_err) = run_post_execute(interceptors, &review_req, &r).await {
-                    tracing::warn!(
-                        agent_round,
-                        error = %val_err,
-                        "post-execute validation failed in agent review; continuing"
-                    );
-                }
-                r
-            }
-            Ok(Err(e)) => {
-                // Quota/billing failures are not retryable — break immediately.
-                // Do NOT activate the global rate-limit circuit breaker: the reviewer
-                // agent is configured independently from the implementation agent and a
-                // depleted reviewer account must not stall unrelated implementation tasks.
-                if matches!(
-                    e,
-                    HarnessError::QuotaExhausted(_) | HarnessError::BillingFailed(_)
-                ) {
-                    tracing::error!(agent_round, error = %e, "quota/billing failure during agent review — aborting");
-                    run_on_error(interceptors, &review_req, &e.to_string()).await;
-                    mutate_and_persist(store, task_id, |s| {
-                        s.status = TaskStatus::Failed;
-                        s.error = Some(e.to_string());
-                    })
-                    .await?;
-                    return Ok((false, false));
-                }
-                run_on_error(interceptors, &review_req, &e.to_string()).await;
-                return Err(e.into());
-            }
-            Err(_) => {
-                let msg = format!(
-                    "Agent review round {agent_round} timed out after {}s",
-                    turn_timeout.as_secs()
-                );
-                run_on_error(interceptors, &review_req, &msg).await;
-                return Err(anyhow::anyhow!("{msg}"));
-            }
-        };
-
-        let AgentResponse { output, stderr, .. } = resp;
-
-        if !stderr.is_empty() {
-            tracing::warn!(agent_round, stderr = %stderr, "agent reviewer stderr");
-        }
-
-        let approved = prompts::is_approved(&output);
-        let issues = prompts::extract_review_issues(&output);
-        let review_detail = output;
-
-        mutate_and_persist(store, task_id, |s| {
-            s.rounds.push(RoundResult {
-                turn: 0, // agent review rounds use turn 0
-                action: "agent_review".into(),
-                result: if approved {
-                    "approved".into()
-                } else {
-                    format!("{} issues", issues.len())
-                },
-                detail: Some(review_detail),
-                first_token_latency_ms: None,
-            });
-        })
-        .await?;
-
-        // Log agent_review event
-        let mut ev = Event::new(
-            SessionId::new(),
-            "agent_review",
-            "task_runner",
-            if approved {
-                harness_core::types::Decision::Complete
-            } else {
-                harness_core::types::Decision::Warn
-            },
-        );
-        ev.detail = Some(format!("pr={pr_url}"));
-        ev.reason = Some(if approved {
-            format!("round {agent_round}: approved")
-        } else {
-            format!("round {agent_round}: {} issues", issues.len())
-        });
-        if let Err(e) = events.log(&ev).await {
-            tracing::warn!("failed to log agent_review event: {e}");
-        }
-
-        if approved {
-            tracing::info!("agent review approved at round {agent_round}");
-            break;
-        }
-
-        // Malformed reviewer output: neither APPROVED nor any ISSUE: lines.
-        // Sending an empty fix prompt would produce arbitrary or no-op commits,
-        // so treat this as a reviewer protocol failure and abort the review loop.
-        if issues.is_empty() {
-            tracing::warn!(
-                agent_round,
-                "agent reviewer output contained neither APPROVED nor ISSUE: lines; \
-                 treating as protocol failure and skipping fix round"
-            );
-            break;
-        }
-
-        // Detect impasse: track how many consecutive rounds produced identical issues.
-        // Must happen before the max_rounds break so thresholds are reachable at the last round.
-        // Compare normalized issue lists directly to avoid false positives from hash collisions.
-        let normalized = normalize_issues(&issues);
-        let consecutive_count = match &impasse_tracker {
-            Some((prev, c)) if *prev == normalized => c + 1,
-            _ => 1,
-        };
-        impasse_tracker = Some((normalized, consecutive_count));
-
-        // 5 consecutive rounds with identical issues → mark task as Failed immediately.
-        if consecutive_count >= 5 {
-            tracing::warn!(
-                agent_round,
-                consecutive_count,
-                "agent review impasse: same issues repeated 5 times — marking task as failed"
-            );
-            mutate_and_persist(store, task_id, |s| {
-                s.status = TaskStatus::Failed;
-                s.error = Some(format!(
-                    "Impasse detected: identical issues repeated {consecutive_count} consecutive rounds."
-                ));
-            })
-            .await?;
-            return Ok((false, false));
-        }
-
-        // 3+ consecutive rounds with identical issues → use the intervention prompt.
-        let is_impasse = consecutive_count >= 3;
-        if is_impasse {
-            tracing::warn!(
-                agent_round,
-                consecutive_count,
-                "agent review impasse detected — same issues repeated, using intervention prompt"
-            );
-        }
-
-        // Skip the fix round only when rounds are exhausted and there is no impasse.
-        // When impasse is detected at the last round, we still apply the intervention prompt
-        // to give the agent one final attempt to break the cycle before GitHub review.
-        if agent_round == max_rounds && !is_impasse {
-            tracing::info!(
-                "agent review exhausted {max_rounds} rounds, proceeding to GitHub review"
-            );
-            break;
-        }
-
-        // Implementor fixes the issues
-        let fix_req = AgentRequest {
-            prompt: {
-                let prompt_fn = if is_impasse {
-                    prompts::agent_review_intervention_prompt
-                } else {
-                    prompts::agent_review_fix_prompt
-                };
-                prompt_fn(pr_url, &issues, agent_round, project_type)
-            },
-            project_root: project.to_path_buf(),
-            context: context_items.to_vec(),
-            execution_phase: Some(ExecutionPhase::Execution),
-            env_vars: cargo_env.clone(),
-            ..Default::default()
-        };
-        let fix_req = run_pre_execute(interceptors, fix_req).await?;
-
-        if let Some(max) = effective_max_turns {
-            if *turns_used >= max {
-                return Err(anyhow::anyhow!(
-                    "Turn budget exhausted before agent review fix round {agent_round}: used {} of {} allowed turns",
-                    turns_used, max
-                ));
-            }
-        }
-        let fix_resp = tokio::time::timeout(turn_timeout, agent.execute(fix_req.clone())).await;
-        *turns_used += 1;
-        match fix_resp {
-            Ok(Ok(r)) => {
-                if let Some(val_err) = run_post_execute(interceptors, &fix_req, &r).await {
-                    tracing::warn!(
-                        agent_round,
-                        error = %val_err,
-                        "post-execute validation failed in agent review fix; continuing"
-                    );
-                }
-            }
-            Ok(Err(e)) => {
-                run_on_error(interceptors, &fix_req, &e.to_string()).await;
-                return Err(e.into());
-            }
-            Err(_) => {
-                let msg = format!(
-                    "Agent review fix round {agent_round} timed out after {}s",
-                    turn_timeout.as_secs()
-                );
-                run_on_error(interceptors, &fix_req, &msg).await;
-                return Err(anyhow::anyhow!("{msg}"));
-            }
-        }
-
-        mutate_and_persist(store, task_id, |s| {
-            s.rounds.push(RoundResult {
-                turn: 0,
-                action: "agent_review_fix".into(),
-                result: "fixed".into(),
-                detail: None,
-                first_token_latency_ms: None,
-            });
-        })
-        .await?;
-        pushed_commit = true;
-    }
-
-    Ok((true, pushed_commit))
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn periodic_review_source_uses_standard_allowed_tools() {
-        // Verifies that periodic_review tasks get a non-empty allowed_tools list,
-        // which causes claude.rs to pass --allowedTools (hard enforcement) instead
-        // of --dangerously-skip-permissions.
-        let tools = restricted_tools(CapabilityProfile::Standard).unwrap_or_default();
-        assert_eq!(
-            tools,
-            CapabilityProfile::Standard.tools().unwrap_or_default()
-        );
-        assert!(!tools.is_empty());
-    }
-
-    #[test]
-    fn standard_implementation_turn_uses_full_profile() {
-        // Non-periodic_review tasks use None → Full profile →
-        // --dangerously-skip-permissions in claude.rs.
-        let implementation_allowed_tools: Option<Vec<String>> = None;
-        assert!(implementation_allowed_tools.is_none());
-    }
-
-    #[test]
-    fn exponential_backoff_three_consecutive_retries() {
-        let base_ms: u64 = 10_000;
-        let max_ms: u64 = 300_000;
-
-        // Attempt 1: base_ms * 2^0 = 10_000 ms (10 s)
-        assert_eq!(compute_backoff_ms(base_ms, max_ms, 1), 10_000);
-        // Attempt 2: base_ms * 2^1 = 20_000 ms (20 s)
-        assert_eq!(compute_backoff_ms(base_ms, max_ms, 2), 20_000);
-        // Attempt 3: base_ms * 2^2 = 40_000 ms (40 s)
-        assert_eq!(compute_backoff_ms(base_ms, max_ms, 3), 40_000);
-    }
-
-    #[test]
-    fn exponential_backoff_capped_at_max() {
-        let base_ms: u64 = 10_000;
-        let max_ms: u64 = 300_000;
-
-        // Attempt 6: 10_000 * 2^5 = 320_000 — should be capped at 300_000
-        assert_eq!(compute_backoff_ms(base_ms, max_ms, 6), 300_000);
-    }
-
-    #[test]
-    fn parse_harness_review_command() {
-        let cmd = parse_harness_mention_command("@harness review");
-        assert_eq!(cmd, Some(HarnessMentionCommand::Review));
-    }
-
-    #[test]
-    fn parse_harness_fix_ci_command_case_insensitive() {
-        let cmd = parse_harness_mention_command("please @Harness FIX CI");
-        assert_eq!(cmd, Some(HarnessMentionCommand::FixCi));
-    }
-
-    #[test]
-    fn parse_harness_plain_mention_command() {
-        let cmd = parse_harness_mention_command("hello @harness can you help?");
-        assert_eq!(cmd, Some(HarnessMentionCommand::Mention));
-    }
-
-    #[test]
-    fn parse_harness_command_returns_none_without_mention() {
-        let cmd = parse_harness_mention_command("no command here");
-        assert_eq!(cmd, None);
-    }
-
-    #[test]
-    fn parse_harness_first_mention_per_line_is_used() {
-        let cmd = parse_harness_mention_command("@harness review then @harness fix ci");
-        assert_eq!(cmd, Some(HarnessMentionCommand::Review));
-    }
-
-    #[test]
-    fn prompt_builder_no_sections_adds_trailing_newline() {
-        let result = PromptBuilder::new("Title line.").build();
-        assert_eq!(result, "Title line.\n");
-    }
-
-    #[test]
-    fn prompt_builder_optional_url_absent_is_skipped() {
-        let result = PromptBuilder::new("Title.")
-            .add_optional_url("Link", None)
-            .build();
-        assert_eq!(result, "Title.\n");
-    }
-
-    #[test]
-    fn prompt_builder_optional_url_present_appears_in_output() {
-        let result = PromptBuilder::new("Title.")
-            .add_optional_url("Link", Some("https://example.com"))
-            .build();
-        assert!(result.contains("- Link: "));
-        assert!(result.contains("https://example.com"));
-        assert!(result.ends_with('\n'));
-    }
-
-    #[test]
-    fn prompt_builder_add_section_wraps_external_data() {
-        let result = PromptBuilder::new("Title.")
-            .add_section("Payload", "content here")
-            .build();
-        assert!(result.contains("Payload:\n"));
-        assert!(result.contains("<external_data>"));
-        assert!(result.contains("content here"));
-    }
-
-    #[test]
-    fn prompt_builder_multiple_urls_all_appear() {
-        let result = PromptBuilder::new("Title.")
-            .add_optional_url("First", Some("url1"))
-            .add_optional_url("Second", None)
-            .add_optional_url("Third", Some("url3"))
-            .build();
-        assert!(result.contains("- First: "));
-        assert!(result.contains("url1"));
-        assert!(!result.contains("Second"));
-        assert!(result.contains("- Third: "));
-        assert!(result.contains("url3"));
-    }
-
-    #[test]
-    fn build_fix_ci_prompt_contains_context() {
-        let prompt = build_fix_ci_prompt(
-            "majiayu000/harness",
-            42,
-            "@harness fix CI",
-            Some("https://github.com/majiayu000/harness/issues/42#issuecomment-1"),
-            Some("https://github.com/majiayu000/harness/pull/42"),
-        );
-
-        assert!(prompt.contains("CI failure repair requested for PR #42"));
-        assert!(prompt.contains("majiayu000/harness"));
-        assert!(prompt.contains("<external_data>"));
-        assert!(prompt.contains("PR_URL=https://github.com/majiayu000/harness/pull/42"));
-    }
-
-    #[test]
-    fn truncate_short_string_passes_through() {
-        let input = "short error";
-        let result = truncate_validation_error(input, 100);
-        assert_eq!(result, "short error");
-    }
-
-    #[test]
-    fn truncate_at_max_chars_boundary() {
-        let input = "a".repeat(200);
-        let result = truncate_validation_error(&input, 50);
-        assert!(result.starts_with(&"a".repeat(50)));
-        assert!(result.contains("(output truncated, 200 chars total)"));
-    }
-
-    #[test]
-    fn truncate_preserves_utf8_boundary() {
-        // "é" is 2 bytes; build a string where max_chars lands mid-character.
-        let input = "ééééé"; // 10 bytes, 5 chars
-        let result = truncate_validation_error(input, 3); // byte 3 is mid-char
-                                                          // Should back up to byte 2 (1 full "é").
-        assert!(result.starts_with("é"));
-        assert!(result.contains("(output truncated,"));
-    }
-
-    #[test]
-    fn review_check_turn_uses_readonly_profile() {
-        let tools = restricted_tools(CapabilityProfile::ReadOnly).unwrap();
-        assert!(tools.contains(&"Read".to_string()));
-        assert!(tools.contains(&"Grep".to_string()));
-        assert!(tools.contains(&"Glob".to_string()));
-        assert!(!tools.contains(&"Write".to_string()));
-        assert!(!tools.contains(&"Edit".to_string()));
-        assert!(!tools.contains(&"Bash".to_string()));
-    }
-
-    #[test]
-    fn periodic_review_turn_uses_standard_profile_with_bash() {
-        let tools = restricted_tools(CapabilityProfile::Standard).unwrap();
-        assert!(tools.contains(&"Bash".to_string()));
-        assert!(tools.contains(&"Read".to_string()));
-        assert!(tools.contains(&"Write".to_string()));
-        assert!(tools.contains(&"Edit".to_string()));
-        // Standard does not include Grep/Glob — it's distinct from ReadOnly.
-        assert!(!tools.contains(&"Grep".to_string()));
-    }
-
-    #[test]
-    fn implementation_turn_uses_full_profile_no_restriction() {
-        // Full profile returns None — no tool restriction is applied to the agent.
-        assert!(CapabilityProfile::Full.tools().is_none());
-    }
-
-    #[test]
-    fn parse_implementation_outcome_prefers_plan_issue() {
-        let output = "PLAN_ISSUE=Plan missed rollback path\nPR_URL=https://github.com/o/r/pull/123";
-        let parsed = parse_implementation_outcome(output);
-        assert_eq!(
-            parsed,
-            ImplementationOutcome::PlanIssue("Plan missed rollback path".to_string())
-        );
-    }
-
-    #[test]
-    fn parse_implementation_outcome_extracts_pr_when_no_plan_issue() {
-        let output = "Done.\nPR_URL=https://github.com/majiayu000/harness/pull/42";
-        let parsed = parse_implementation_outcome(output);
-        assert_eq!(
-            parsed,
-            ImplementationOutcome::ParsedPr {
-                pr_url: Some("https://github.com/majiayu000/harness/pull/42".to_string()),
-                pr_num: Some(42),
-                created_issue_num: None,
-            }
-        );
-    }
-
-    #[test]
-    fn constitution_present_when_enabled() {
-        let result = prepend_constitution("Do the task.".to_string(), true);
-        assert!(result.contains("GP-01"));
-        assert!(result.contains("GP-02"));
-        assert!(result.contains("GP-03"));
-        assert!(result.contains("GP-04"));
-        assert!(result.contains("GP-05"));
-        assert!(result.ends_with("Do the task."));
-    }
-
-    #[test]
-    fn constitution_absent_when_disabled() {
-        let result = prepend_constitution("Do the task.".to_string(), false);
-        assert_eq!(result, "Do the task.");
-        assert!(!result.contains("GP-01"));
-    }
-
-    #[test]
-    fn normalize_issues_is_order_invariant() {
-        let ordered = vec!["issue A".to_string(), "issue B".to_string()];
-        let reversed = vec!["issue B".to_string(), "issue A".to_string()];
-        assert_eq!(normalize_issues(&ordered), normalize_issues(&reversed));
-    }
-
-    fn step_tracker(
-        tracker: &mut Option<(Vec<String>, u32)>,
-        issues: &[String],
-    ) -> (u32, bool, bool) {
-        let normalized = normalize_issues(issues);
-        let count = match tracker.as_ref() {
-            Some((prev, c)) if *prev == normalized => c + 1,
-            _ => 1,
-        };
-        *tracker = Some((normalized, count));
-        let intervention = count >= 3;
-        let fatal = count >= 5;
-        (count, intervention, fatal)
-    }
-
-    #[test]
-    fn impasse_no_intervention_for_first_two_rounds() {
-        let issues = vec!["null pointer".to_string()];
-        let mut tracker: Option<(Vec<String>, u32)> = None;
-        let (c1, i1, f1) = step_tracker(&mut tracker, &issues);
-        let (c2, i2, f2) = step_tracker(&mut tracker, &issues);
-        assert_eq!(c1, 1);
-        assert!(!i1 && !f1, "no action on first occurrence");
-        assert_eq!(c2, 2);
-        assert!(!i2 && !f2, "no action on second occurrence");
-    }
-
-    #[test]
-    fn impasse_intervention_at_third_consecutive_round() {
-        let issues = vec!["null pointer".to_string()];
-        let mut tracker: Option<(Vec<String>, u32)> = None;
-        step_tracker(&mut tracker, &issues); // round 1
-        step_tracker(&mut tracker, &issues); // round 2
-        let (c3, i3, f3) = step_tracker(&mut tracker, &issues); // round 3
-        assert_eq!(c3, 3);
-        assert!(i3, "intervention at 3rd consecutive round");
-        assert!(!f3, "not yet fatal at round 3");
-    }
-
-    #[test]
-    fn impasse_fatal_at_fifth_consecutive_round() {
-        let issues = vec!["null pointer".to_string()];
-        let mut tracker: Option<(Vec<String>, u32)> = None;
-        for _ in 0..4 {
-            step_tracker(&mut tracker, &issues);
-        }
-        let (c5, i5, f5) = step_tracker(&mut tracker, &issues);
-        assert_eq!(c5, 5);
-        assert!(i5, "intervention still active at round 5");
-        assert!(f5, "fatal at 5th consecutive round");
-    }
-
-    #[test]
-    fn impasse_counter_resets_when_issues_change() {
-        let issues = vec!["null pointer".to_string()];
-        let other = vec!["different bug".to_string()];
-        let mut tracker: Option<(Vec<String>, u32)> = None;
-        step_tracker(&mut tracker, &issues); // count 1
-        step_tracker(&mut tracker, &issues); // count 2
-        step_tracker(&mut tracker, &issues); // count 3 — would trigger intervention
-        let (c_reset, i_reset, _) = step_tracker(&mut tracker, &other); // different issues
-        assert_eq!(c_reset, 1, "counter resets on different issues");
-        assert!(!i_reset, "no intervention after reset");
-    }
-
-    // --- jaccard_word_similarity unit tests ---
-
-    #[test]
-    fn jaccard_identical_strings() {
-        assert_eq!(jaccard_word_similarity("hello world", "hello world"), 1.0);
-    }
-
-    #[test]
-    fn jaccard_disjoint_strings() {
-        assert_eq!(jaccard_word_similarity("foo bar", "baz qux"), 0.0);
-    }
-
-    #[test]
-    fn jaccard_partial_overlap() {
-        // {"a", "b"} ∩ {"b", "c"} = {"b"}, union = {"a","b","c"} → 1/3
-        let score = jaccard_word_similarity("a b", "b c");
-        let expected = 1.0_f64 / 3.0_f64;
-        assert!(
-            (score - expected).abs() < 1e-10,
-            "expected ~{expected}, got {score}"
-        );
-    }
-
-    #[test]
-    fn jaccard_one_empty() {
-        assert_eq!(jaccard_word_similarity("", "hello world"), 0.0);
-        assert_eq!(jaccard_word_similarity("hello world", ""), 0.0);
-    }
-
-    #[test]
-    fn jaccard_both_empty() {
-        assert_eq!(jaccard_word_similarity("", ""), 1.0);
-    }
-}
